gstreamer::caps

Struct Builder

pub struct Builder<T> {
    s: Structure,
    features: Option<CapsFeatures>,
    phantom: PhantomData<T>,
}

Fields
s: Structurefeatures: Option<CapsFeatures>phantom: PhantomData<T>
Implementations

impl Builder<NoFeature>

fn new(name: impl IntoGStr) -> Builder<NoFeature>

fn from_static(name: impl AsRef<GStr> + 'static) -> Builder<NoFeature>

fn from_id(name: impl AsRef<IdStr>) -> Builder<NoFeature>

pub fn features<S: IntoGStr>( self, features: impl IntoIterator<Item = S>, ) -> Builder<HasFeatures>

pub fn features_from_statics<S: AsRef<GStr> + 'static>( self, features: impl IntoIterator<Item = S>, ) -> Builder<HasFeatures>

pub fn features_from_ids<S: AsRef<IdStr>>( self, features: impl IntoIterator<Item = S>, ) -> Builder<HasFeatures>

pub fn any_features(self) -> Builder<HasFeatures>

impl<T> Builder<T>

pub fn field(self, name: impl IntoGStr, value: impl Into<Value> + Send) -> Self

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn field_with_static( self, name: impl AsRef<GStr> + 'static, value: impl Into<Value> + Send, ) -> Self

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn field_with_id( self, name: impl AsRef<IdStr>, value: impl Into<Value> + Send, ) -> Self

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn field_if( self, name: impl IntoGStr, value: impl Into<Value> + Send, predicate: bool, ) -> Self

Sets field name to the given inner value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn field_with_static_if( self, name: impl AsRef<GStr> + 'static, value: impl Into<Value> + Send, predicate: bool, ) -> Self

Sets field name to the given inner value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn field_with_id_if( self, name: impl AsRef<IdStr>, value: impl Into<Value> + Send, predicate: bool, ) -> Self

Sets field name to the given inner value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn field_if_some( self, name: impl IntoGStr, value: Option<impl Into<Value> + Send>, ) -> Self

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn field_with_static_if_some( self, name: impl AsRef<GStr> + 'static, value: Option<impl Into<Value> + Send>, ) -> Self

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn field_with_id_if_some( self, name: impl AsRef<IdStr>, value: Option<impl Into<Value> + Send>, ) -> Self

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn field_from_iter<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl IntoGStr, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn field_with_static_from_iter<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl AsRef<GStr> + 'static, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn field_with_id_from_iter<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl AsRef<IdStr>, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn field_if_not_empty<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl IntoGStr, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn field_with_static_if_not_empty<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl AsRef<GStr> + 'static, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn field_with_id_if_not_empty<V: ValueType + Into<Value> + FromIterator<SendValue> + Send, I: ToSendValue>( self, name: impl AsRef<IdStr>, iter: impl IntoIterator<Item = I>, ) -> Self

Sets field name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn build(self) -> Caps

pub fn structure(&self) -> &Structure