gstreamer::caps

Struct CapsRef

#[repr(transparent)]
pub struct CapsRef(GstCaps);

Tuple Fields
0: GstCaps
Implementations

impl CapsRef

pub fn as_ptr(&self) -> *const GstCaps

pub fn as_mut_ptr(&self) -> *mut GstCaps

pub unsafe fn from_ptr<'a>(ptr: *const GstCaps) -> &'a Self

pub unsafe fn from_mut_ptr<'a>(ptr: *mut GstCaps) -> &'a mut Self

pub fn copy(&self) -> Caps

pub fn upcast_ref(&self) -> &MiniObjectRef

pub fn upcast_mut(&mut self) -> &mut MiniObjectRef

pub fn ptr_eq(this: &CapsRef, other: &CapsRef) -> bool

pub fn mark_may_be_leaked(&mut self)

impl CapsRef

pub fn set(&mut self, name: impl IntoGStr, value: impl ToSendValue + Sync)

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn set_with_static( &mut self, name: impl AsRef<GStr> + 'static, value: impl ToSendValue + Sync, )

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn set_with_id( &mut self, name: impl AsRef<IdStr>, value: impl ToSendValue + Sync, )

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn set_if( &mut self, name: impl IntoGStr, value: impl ToSendValue + Sync, predicate: bool, )

Sets field name to the given value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn set_with_static_if( &mut self, name: impl AsRef<GStr> + 'static, value: impl ToSendValue + Sync, predicate: bool, )

Sets field name to the given value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn set_with_id_if( &mut self, name: impl AsRef<IdStr>, value: impl ToSendValue + Sync, predicate: bool, )

Sets field name to the given value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn set_if_some( &mut self, name: impl IntoGStr, value: Option<impl ToSendValue + Sync>, )

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn set_with_static_if_some( &mut self, name: impl AsRef<GStr> + 'static, value: Option<impl ToSendValue + Sync>, )

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn set_with_id_if_some( &mut self, name: impl AsRef<IdStr>, value: Option<impl ToSendValue + Sync>, )

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn set_from_iter<V: ValueType + ToSendValue + FromIterator<SendValue> + Sync, I: ToSendValue>( &mut self, name: impl IntoGStr, iter: impl IntoIterator<Item = I>, )

Sets field name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn set_with_static_from_iter<V: ValueType + ToSendValue + FromIterator<SendValue> + Sync, I: ToSendValue>( &mut self, name: impl AsRef<GStr> + 'static, iter: impl IntoIterator<Item = I>, )

Sets field name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn set_with_id_from_iter<V: ValueType + ToSendValue + FromIterator<SendValue> + Sync, I: ToSendValue>( &mut self, name: impl AsRef<IdStr>, iter: impl IntoIterator<Item = I>, )

Sets field name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn set_if_not_empty<V: ValueType + ToSendValue + FromIterator<SendValue> + Sync, I: ToSendValue>( &mut self, name: impl IntoGStr, iter: impl IntoIterator<Item = I>, )

Sets field name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn set_with_static_if_not_empty<V: ValueType + ToSendValue + FromIterator<SendValue> + Sync, I: ToSendValue>( &mut self, name: impl AsRef<GStr> + 'static, iter: impl IntoIterator<Item = I>, )

Sets field name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn set_with_id_if_not_empty<V: ValueType + ToSendValue + FromIterator<SendValue> + Sync, I: ToSendValue>( &mut self, name: impl AsRef<IdStr>, iter: impl IntoIterator<Item = I>, )

Sets field name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn set_value(&mut self, name: impl IntoGStr, value: SendValue)

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn set_value_with_static( &mut self, name: impl AsRef<GStr> + 'static, value: SendValue, )

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn set_value_with_id(&mut self, name: impl AsRef<IdStr>, value: SendValue)

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn set_value_if( &mut self, name: impl IntoGStr, value: SendValue, predicate: bool, )

Sets field name to the given value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn set_value_with_static_if( &mut self, name: impl AsRef<GStr> + 'static, value: SendValue, predicate: bool, )

Sets field name to the given value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn set_value_with_id_if( &mut self, name: impl AsRef<IdStr>, value: SendValue, predicate: bool, )

Sets field name to the given value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn set_value_if_some( &mut self, name: impl IntoGStr, value: Option<SendValue>, )

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn set_value_with_static_if_some( &mut self, name: impl AsRef<GStr> + 'static, value: Option<SendValue>, )

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn set_value_with_id_if_some( &mut self, name: impl AsRef<IdStr>, value: Option<SendValue>, )

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn structure(&self, idx: usize) -> Option<&StructureRef>

pub fn structure_mut(&mut self, idx: usize) -> Option<&mut StructureRef>

pub fn features(&self, idx: usize) -> Option<&CapsFeaturesRef>

pub fn features_mut(&mut self, idx: usize) -> Option<&mut CapsFeaturesRef>

pub fn set_features(&mut self, idx: usize, features: Option<CapsFeatures>)

pub fn set_features_simple(&mut self, features: Option<CapsFeatures>)

pub fn size(&self) -> usize

pub fn len(&self) -> usize

pub fn iter(&self) -> Iter<'_>

pub fn iter_mut(&mut self) -> IterMut<'_>

pub fn iter_with_features(&self) -> IterFeatures<'_>

pub fn iter_with_features_mut(&mut self) -> IterFeaturesMut<'_>

pub fn append_structure(&mut self, structure: Structure)

pub fn append_structure_full( &mut self, structure: Structure, features: Option<CapsFeatures>, )

pub fn remove_structure(&mut self, idx: usize)

pub fn append(&mut self, other: Caps)

pub fn can_intersect(&self, other: &Self) -> bool

pub fn intersect(&self, other: &Self) -> Caps

pub fn intersect_with_mode(&self, other: &Self, mode: CapsIntersectMode) -> Caps

pub fn is_always_compatible(&self, other: &Self) -> bool

pub fn is_any(&self) -> bool

pub fn is_empty(&self) -> bool

pub fn is_fixed(&self) -> bool

pub fn is_equal_fixed(&self, other: &Self) -> bool

pub fn is_strictly_equal(&self, other: &Self) -> bool

pub fn is_subset(&self, superset: &Self) -> bool

pub fn is_subset_structure(&self, structure: &StructureRef) -> bool

pub fn is_subset_structure_full( &self, structure: &StructureRef, features: Option<&CapsFeaturesRef>, ) -> bool

pub fn subtract(&self, other: &Self) -> Caps

pub fn serialize(&self, flags: SerializeFlags) -> GString

pub fn foreach<F: FnMut(&CapsFeaturesRef, &StructureRef) -> ControlFlow<()>>( &self, func: F, ) -> bool

pub fn map_in_place<F: FnMut(&mut CapsFeaturesRef, &mut StructureRef) -> ControlFlow<()>>( &mut self, func: F, )

pub fn filter_map_in_place<F: FnMut(&mut CapsFeaturesRef, &mut StructureRef) -> CapsFilterMapAction>( &mut self, func: F, )