Enum gstreamer::EventType

pub enum EventType {
    Unknown,
    FlushStart,
    FlushStop,
    StreamStart,
    Caps,
    Segment,
    StreamCollection,
    Tag,
    Buffersize,
    SinkMessage,
    StreamGroupDone,
    Eos,
    Toc,
    Protection,
    SegmentDone,
    Gap,
    Qos,
    Seek,
    Navigation,
    Latency,
    Step,
    Reconfigure,
    TocSelect,
    SelectStreams,
    CustomUpstream,
    CustomDownstream,
    CustomDownstreamOob,
    CustomDownstreamSticky,
    CustomBoth,
    CustomBothOob,
    // some variants omitted
}

EventType lists the standard event types that can be sent in a pipeline.

The custom event types can be used for private messages between elements that can't be expressed using normal GStreamer buffer passing semantics. Custom events carry an arbitrary Structure. Specific custom events are distinguished by the name of the structure.

Variants
UnknownFlushStartFlushStopStreamStartCapsSegmentStreamCollectionTagBuffersizeSinkMessageStreamGroupDoneEosTocProtectionSegmentDoneGapQosSeekNavigationLatencyStepReconfigureTocSelectSelectStreamsCustomUpstreamCustomDownstreamCustomDownstreamOobCustomDownstreamStickyCustomBothCustomBothOob
Methods

impl EventType

pub fn is_upstream(&self) -> bool

pub fn is_downstream(&self) -> bool

pub fn is_serialized(&self) -> bool

pub fn is_sticky(&self) -> bool

pub fn is_sticky_multi(&self) -> bool