gstreamer::event

Struct StreamStartBuilder

pub struct StreamStartBuilder<'a> { /* private fields */ }

Implementations

impl<'a> StreamStartBuilder<'a>

pub fn flags(self, flags: StreamFlags) -> Self

pub fn flags_if(self, flags: StreamFlags, predicate: bool) -> Self

pub fn flags_if_some(self, flags: Option<StreamFlags>) -> Self

pub fn group_id(self, group_id: GroupId) -> Self

pub fn group_id_if(self, group_id: GroupId, predicate: bool) -> Self

pub fn group_id_if_some(self, group_id: Option<GroupId>) -> Self

pub fn stream(self, stream: Stream) -> Self

pub fn stream_if(self, stream: Stream, predicate: bool) -> Self

pub fn stream_if_some(self, stream: Option<Stream>) -> Self

pub fn seqnum(self, seqnum: Seqnum) -> Self

pub fn seqnum_if(self, seqnum: Seqnum, predicate: bool) -> Self

pub fn seqnum_if_some(self, seqnum: Option<Seqnum>) -> Self

pub fn running_time_offset(self, running_time_offset: i64) -> Self

pub fn running_time_offset_if( self, running_time_offset: i64, predicate: bool, ) -> Self

pub fn running_time_offset_if_some( self, running_time_offset: Option<i64>, ) -> Self

pub fn other_field(self, name: &'a str, value: impl ToSendValue) -> Self

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn other_field_if( self, name: &'a str, value: impl ToSendValue, predicate: bool, ) -> Self

Sets field name to the given inner value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn other_field_if_some( self, name: &'a str, value: Option<impl ToSendValue>, ) -> Self

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn other_field_from_iter<V: ValueType + ToSendValue + FromIterator<SendValue>>( self, name: &'a str, iter: impl IntoIterator<Item = impl ToSendValue>, ) -> Self

Sets field name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn other_field_if_not_empty<V: ValueType + ToSendValue + FromIterator<SendValue>>( self, name: &'a str, iter: impl IntoIterator<Item = impl ToSendValue>, ) -> Self

Sets field name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn other_fields( self, other_fields: &[(&'a str, &'a (dyn ToSendValue + Sync))], ) -> Self

👎Deprecated: use build.other_field() instead

pub fn build(self) -> Event