gstreamer::format

Module clock_time

Structs

ClockTime

A Time quantity

DisplayableOptClockTime

DurationError

TryFromFloatSecsError

Enums

ClockTimeValueTypeOrNoneChecker

Sign 🔒

Tell pad_clocktime what kind of time we’re formatting

Constants

TRY_FROM_FLOAT_SECS_ERROR_MSG 🔒

Traits

TimeFormatConstructor

ClockTime formatted value constructor trait.

Functions

fmt_opt_clock_time 🔒

pad_clocktime 🔒

Performs the correct padding for a clock time which has already been emitted into a str, as by write_clocktime. The str should not contain the sign; that will be added by this method.

write_clocktime 🔒

Writes an unpadded, signless clocktime string with the given precision