gstreamer::format::clock_time

Trait TimeFormatConstructor

pub trait TimeFormatConstructor {
    // Required methods
    fn nseconds(self) -> ClockTime;
    fn useconds(self) -> ClockTime;
    fn mseconds(self) -> ClockTime;
    fn seconds(self) -> ClockTime;
    fn minutes(self) -> ClockTime;
    fn hours(self) -> ClockTime;
}

ClockTime formatted value constructor trait.

Required Methods

fn nseconds(self) -> ClockTime

Builds a ClockTime formatted value from self interpreted as nano seconds.

fn useconds(self) -> ClockTime

Builds a ClockTime formatted value from self interpreted as micro seconds.

fn mseconds(self) -> ClockTime

Builds a ClockTime formatted value from self interpreted as milli seconds.

fn seconds(self) -> ClockTime

Builds a ClockTime formatted value from self interpreted as seconds.

fn minutes(self) -> ClockTime

Builds a ClockTime formatted value from self interpreted as minutes.

fn hours(self) -> ClockTime

Builds a ClockTime formatted value from self interpreted as hours.