Trait gstreamer::format::signed::NoneSignedBuilder

pub trait NoneSignedBuilder: FormattedValueNoneBuilder {
    type Signed;

    // Required methods
    fn none_signed() -> Self::Signed;
    fn none_signed_for_format(format: Format) -> Self::Signed;
}

Required Associated Types

type Signed

Required Methods

fn none_signed() -> Self::Signed

Returns the None value for Self as a Signed<FullRange> if such a value can be represented.

See details in FormattedValueNoneBuilder::none.

Panics

Panics if Self is GenericFormattedValue in which case, the Format must be known.

fn none_signed_for_format(format: Format) -> Self::Signed

Returns the None value for Self as a Signed<FullRange>, if such a value can be represented.

See details in FormattedValueNoneBuilder::none_for_format.

Panics

Panics if None can’t be represented by Self for format or by the requested GenericFormattedValue variant.