gstreamer::format::specific

Struct Default

pub struct Default(u64);

A unit-less quantity

Some functions enforce format specific quantities. This type can be used when a Default format is expected. It comes with functions to perform computations without the need to retrieve the inner integer.

Examples

// Regular constructors (can be used in `const` contexts)
const FORTY_TWO: Default = Default::from_u64(42);
let two = Default::from_u64(2);

// All four arithmetic operations
let limit = (FORTY_TWO + two) * 2 / 3;

// Comparisons
if limit > Default::ZERO {
    println!("Greater");
}

See the documentation of the format module for more examples.

Tuple Fields
0: u64
Implementations

impl Default

pub const MAX: Self

impl Default

pub const fn from_u64(quantity: u64) -> Self

Builds a new Default formatted value with the provided quantity.

Panics

Panics if the provided quantity equals u64::MAX, which is reserved for None in C.

pub fn from_usize(quantity: usize) -> Self

Builds a new Default formatted value with the provided quantity.

Panics

Panics if the provided quantity equals u64::MAX, which is reserved for None in C.

impl Default

pub const ZERO: Self

pub const NONE: Option<Self> = None

pub const ONE: Self

The unitary value.

pub const MAX_SIGNED: Signed<Default>

pub const MIN_SIGNED: Signed<Default>

pub const fn is_zero(self) -> bool

impl Default

pub const fn checked_add(self, rhs: Self) -> Option<Self>

pub const fn saturating_add(self, rhs: Self) -> Self

pub fn overflowing_add(self, rhs: Self) -> (Self, bool)

pub fn wrapping_add(self, rhs: Self) -> Self

pub const fn checked_sub(self, rhs: Self) -> Option<Self>

pub const fn saturating_sub(self, rhs: Self) -> Self

pub const fn overflowing_sub(self, rhs: Self) -> (Self, bool)

pub const fn wrapping_sub(self, rhs: Self) -> Self

pub fn absdiff(self, rhs: Self) -> Self

impl Default

pub const fn checked_div(self, rhs: u64) -> Option<Self>

pub const fn saturating_div(self, rhs: u64) -> Self

pub const fn checked_mul(self, rhs: u64) -> Option<Self>

pub const fn saturating_mul(self, rhs: u64) -> Self

pub fn overflowing_mul(self, rhs: u64) -> (Self, bool)

pub fn wrapping_mul(self, rhs: u64) -> Self

pub const fn checked_rem(self, rhs: u64) -> Option<Self>