Struct gstreamer::format::Buffers

pub struct Buffers(u64);

Tuple Fields
0: u64
Implementations

impl Buffers

pub const OFFSET_NONE: u64 = 18_446_744_073_709_551_615u64

pub const MAX: Self = _

impl Buffers

pub const fn from_u64(buffers: u64) -> Self

Builds a new Buffers formatted value with the provided buffers count.

Panics

Panics if the provided count equals u64::MAX, which is reserved for None in C.

pub fn from_usize(buffers: usize) -> Self

Builds a new Buffers formatted value with the provided buffers count.

Panics

Panics if the provided count equals u64::MAX, which is reserved for None in C.

impl Buffers

pub const ZERO: Self = _

pub const NONE: Option<Self> = None

pub const ONE: Self = _

The unitary value.

pub const MAX_SIGNED: Signed<Buffers> = _

pub const MIN_SIGNED: Signed<Buffers> = _

pub const fn is_zero(self) -> bool

impl Buffers

pub const fn checked_add(self, rhs: Self) -> Option<Self>

pub const fn saturating_add(self, rhs: Self) -> Self

pub fn overflowing_add(self, rhs: Self) -> (Self, bool)

pub fn wrapping_add(self, rhs: Self) -> Self

pub const fn checked_sub(self, rhs: Self) -> Option<Self>

pub const fn saturating_sub(self, rhs: Self) -> Self

pub const fn overflowing_sub(self, rhs: Self) -> (Self, bool)

pub const fn wrapping_sub(self, rhs: Self) -> Self

pub fn absdiff(self, rhs: Self) -> Self

impl Buffers

pub const fn checked_div(self, rhs: u64) -> Option<Self>

pub const fn saturating_div(self, rhs: u64) -> Self

pub const fn checked_mul(self, rhs: u64) -> Option<Self>

pub const fn saturating_mul(self, rhs: u64) -> Self

pub fn overflowing_mul(self, rhs: u64) -> (Self, bool)

pub fn wrapping_mul(self, rhs: u64) -> Self

pub const fn checked_rem(self, rhs: u64) -> Option<Self>