gstreamer::format

Struct Other

pub struct Other(/* private fields */);

Implementations

impl Other

pub const MAX: Self

impl Other

pub const fn from_u64(quantity: u64) -> Self

Builds a new Other value with the provided quantity.

Panics

Panics if the provided quantity equals u64::MAX, which is reserved for None in C.

pub fn from_usize(quantity: usize) -> Self

Builds a new Other value with the provided quantity.

Panics

Panics if the provided quantity equals u64::MAX, which is reserved for None in C.

impl Other

pub const ZERO: Self

pub const NONE: Option<Self> = None

pub const ONE: Self

The unitary value.

pub const MAX_SIGNED: Signed<Other>

pub const MIN_SIGNED: Signed<Other>

pub const fn is_zero(self) -> bool

impl Other

pub const fn checked_add(self, rhs: Self) -> Option<Self>

pub const fn saturating_add(self, rhs: Self) -> Self

pub fn overflowing_add(self, rhs: Self) -> (Self, bool)

pub fn wrapping_add(self, rhs: Self) -> Self

pub const fn checked_sub(self, rhs: Self) -> Option<Self>

pub const fn saturating_sub(self, rhs: Self) -> Self

pub const fn overflowing_sub(self, rhs: Self) -> (Self, bool)

pub const fn wrapping_sub(self, rhs: Self) -> Self

pub fn absdiff(self, rhs: Self) -> Self

impl Other

pub const fn checked_div(self, rhs: u64) -> Option<Self>

pub const fn saturating_div(self, rhs: u64) -> Self

pub const fn checked_mul(self, rhs: u64) -> Option<Self>

pub const fn saturating_mul(self, rhs: u64) -> Self

pub fn overflowing_mul(self, rhs: u64) -> (Self, bool)

pub fn wrapping_mul(self, rhs: u64) -> Self

pub const fn checked_rem(self, rhs: u64) -> Option<Self>