Type Alias gstreamer::format::Time

pub type Time = ClockTime;

Aliased Type

struct Time(u64);

Fields
0: u64
Implementations

impl ClockTime

pub const SECOND: ClockTime = _

pub const MSECOND: ClockTime = _

pub const USECOND: ClockTime = _

pub const NSECOND: ClockTime = _

pub const MAX: ClockTime = _

pub const fn hours(self) -> u64

pub const fn minutes(self) -> u64

pub const fn seconds(self) -> u64

pub fn seconds_f32(self) -> f32

pub fn seconds_f64(self) -> f64

pub const fn mseconds(self) -> u64

pub const fn useconds(self) -> u64

pub const fn nseconds(self) -> u64

pub const fn from_seconds(seconds: u64) -> Self

Builds a new ClockTime which value is the given number of seconds.

Panics

Panics if the resulting duration in nanoseconds exceeds the u64 range.

pub fn try_from_seconds_f32(seconds: f32) -> Result<Self, TryFromFloatSecsError>

Builds a new ClockTime which value is the given number of seconds.

Returns an error if seconds is negative, infinite or NaN, or the resulting duration in nanoseconds exceeds the u64 range.

pub fn from_seconds_f32(seconds: f32) -> Self

Builds a new ClockTime which value is the given number of seconds.

Panics

Panics if seconds is negative, infinite or NaN, or the resulting duration in nanoseconds exceeds the u64 range.

pub fn try_from_seconds_f64(seconds: f64) -> Result<Self, TryFromFloatSecsError>

Builds a new ClockTime which value is the given number of seconds.

Returns an error if seconds is negative, infinite or NaN, or the resulting duration in nanoseconds exceeds the u64 range.

pub fn from_seconds_f64(seconds: f64) -> Self

Builds a new ClockTime which value is the given number of seconds.

Panics

Panics if seconds is negative, infinite or NaN, or the resulting duration in nanoseconds exceeds the u64 range.

pub const fn from_mseconds(mseconds: u64) -> Self

Builds a new ClockTime which value is the given number of milliseconds.

Panics

Panics if the resulting duration in nanoseconds exceeds the u64 range.

pub const fn from_useconds(useconds: u64) -> Self

Builds a new ClockTime which value is the given number of microseconds.

Panics

Panics if the resulting duration in nanoseconds exceeds the u64 range.

pub const fn from_nseconds(nseconds: u64) -> Self

Builds a new ClockTime which value is the given number of nanoseconds.

Panics

Panics if the requested duration equals GST_CLOCK_TIME_NONE (u64::MAX).

impl ClockTime

pub const ZERO: Self = _

pub const NONE: Option<Self> = None

pub const ONE: Self = _

The unitary value.

pub const MAX_SIGNED: Signed<ClockTime> = _

pub const MIN_SIGNED: Signed<ClockTime> = _

pub const fn is_zero(self) -> bool

impl ClockTime

pub const fn checked_add(self, rhs: Self) -> Option<Self>

pub const fn saturating_add(self, rhs: Self) -> Self

pub fn overflowing_add(self, rhs: Self) -> (Self, bool)

pub fn wrapping_add(self, rhs: Self) -> Self

pub const fn checked_sub(self, rhs: Self) -> Option<Self>

pub const fn saturating_sub(self, rhs: Self) -> Self

pub const fn overflowing_sub(self, rhs: Self) -> (Self, bool)

pub const fn wrapping_sub(self, rhs: Self) -> Self

pub fn absdiff(self, rhs: Self) -> Self

impl ClockTime

pub const fn checked_div(self, rhs: u64) -> Option<Self>

pub const fn saturating_div(self, rhs: u64) -> Self

pub const fn checked_mul(self, rhs: u64) -> Option<Self>

pub const fn saturating_mul(self, rhs: u64) -> Self

pub fn overflowing_mul(self, rhs: u64) -> (Self, bool)

pub fn wrapping_mul(self, rhs: u64) -> Self

pub const fn checked_rem(self, rhs: u64) -> Option<Self>