Struct gstreamer::miniobject::GstRc

pub struct GstRc<T: MiniObject> { /* fields omitted */ }

Methods

impl<T: MiniObject> GstRc<T>

unsafe fn from_glib_none(ptr: *const T::GstType) -> Self

unsafe fn from_glib_full(ptr: *const T::GstType) -> Self

unsafe fn from_glib_borrow(ptr: *const T::GstType) -> Self

fn make_mut(&mut self) -> &mut T

fn get_mut(&mut self) -> Option<&mut T>

fn copy(&self) -> Self

fn is_writable(&self) -> bool

unsafe fn into_ptr(self) -> *mut T::GstType

impl GstRc<MessageRef>

fn new_eos() -> EosBuilder

Create a new eos message. This message is generated and posted in the sink elements of a GstBin. The bin will only forward the EOS message to the application if all sinks have posted an EOS message.

src

The object originating the message.

Returns

The new eos message.

MT safe.

fn new_error(error: &Error) -> ErrorBuilder

Create a new error message. The message will copy error and debug. This message is posted by element when a fatal event occurred. The pipeline will probably (partially) stop. The application receiving this message should stop the pipeline.

src

The object originating the message.

error

The GError for this message.

debug

A debugging string.

Returns

the new error message.

MT safe.

fn new_warning(error: &Error) -> WarningBuilder

Create a new warning message. The message will make copies of error and debug.

src

The object originating the message.

error

The GError for this message.

debug

A debugging string.

Returns

the new warning message.

MT safe.

fn new_info(error: &Error) -> InfoBuilder

Create a new info message. The message will make copies of error and debug.

src

The object originating the message.

error

The GError for this message.

debug

A debugging string.

Returns

the new info message.

MT safe.

fn new_tag(tags: &TagList) -> TagBuilder

Create a new tag message. The message will take ownership of the tag list. The message is posted by elements that discovered a new taglist.

src

The object originating the message.

tag_list

the tag list for the message.

Returns

the new tag message.

MT safe.

fn new_buffering(percent: i32) -> BufferingBuilder

Create a new buffering message. This message can be posted by an element that needs to buffer data before it can continue processing. percent should be a value between 0 and 100. A value of 100 means that the buffering completed.

When percent is < 100 the application should PAUSE a PLAYING pipeline. When percent is 100, the application can set the pipeline (back) to PLAYING. The application must be prepared to receive BUFFERING messages in the PREROLLING state and may only set the pipeline to PLAYING after receiving a message with percent set to 100, which can happen after the pipeline completed prerolling.

MT safe.

src

The object originating the message.

percent

The buffering percent

Returns

The new buffering message.

fn new_state_changed(
old: State,
new: State,
pending: State
) -> StateChangedBuilder

Create a state change message. This message is posted whenever an element changed its state.

src

The object originating the message.

oldstate

the previous state

newstate

the new (current) state

pending

the pending (target) state

Returns

the new state change message.

MT safe.

fn new_state_dirty() -> StateDirtyBuilder

Create a state dirty message. This message is posted whenever an element changed its state asynchronously and is used internally to update the states of container objects.

src

The object originating the message

Returns

the new state dirty message.

MT safe.

fn new_step_done(
format: Format,
amount: u64,
rate: f64,
flush: bool,
intermediate: bool,
duration: u64,
eos: bool
) -> StepDoneBuilder

This message is posted by elements when they complete a part, when intermediate set to true, or a complete step operation.

duration will contain the amount of time (in GST_FORMAT_TIME) of the stepped amount of media in format format.

src

The object originating the message.

format

the format of amount

amount

the amount of stepped data

rate

the rate of the stepped amount

flush

is this an flushing step

intermediate

is this an intermediate step

duration

the duration of the data

eos

the step caused EOS

Returns

the new step_done message.

MT safe.

fn new_clock_provide(clock: &Clock, ready: bool) -> ClockProvideBuilder

Create a clock provide message. This message is posted whenever an element is ready to provide a clock or lost its ability to provide a clock (maybe because it paused or became EOS).

This message is mainly used internally to manage the clock selection.

src

The object originating the message.

clock

the clock it provides

ready

true if the sender can provide a clock

Returns

the new provide clock message.

MT safe.

fn new_clock_lost(clock: &Clock) -> ClockLostBuilder

Create a clock lost message. This message is posted whenever the clock is not valid anymore.

If this message is posted by the pipeline, the pipeline will select a new clock again when it goes to PLAYING. It might therefore be needed to set the pipeline to PAUSED and PLAYING again.

src

The object originating the message.

clock

the clock that was lost

Returns

The new clock lost message.

MT safe.

fn new_new_clock(clock: &Clock) -> NewClockBuilder

Create a new clock message. This message is posted whenever the pipeline selects a new clock for the pipeline.

src

The object originating the message.

clock

the new selected clock

Returns

The new new clock message.

MT safe.

fn new_structure_change(
type_: StructureChangeType,
owner: &Element,
busy: bool
) -> StructureChangeBuilder

Create a new structure change message. This message is posted when the structure of a pipeline is in the process of being changed, for example when pads are linked or unlinked.

src should be the sinkpad that unlinked or linked.

src

The object originating the message.

type_

The change type.

owner

The owner element of src.

busy

Whether the structure change is busy.

Returns

the new structure change message.

MT safe.

fn new_stream_status(
type_: StreamStatusType,
owner: &Element
) -> StreamStatusBuilder

Create a new stream status message. This message is posted when a streaming thread is created/destroyed or when the state changed.

src

The object originating the message.

type_

The stream status type.

owner

the owner element of src.

Returns

the new stream status message.

MT safe.

fn new_application(structure: Structure) -> ApplicationBuilder

Create a new application-typed message. GStreamer will never create these messages; they are a gift from us to you. Enjoy.

src

The object originating the message.

structure

the structure for the message. The message will take ownership of the structure.

Returns

The new application message.

MT safe.

fn new_element(structure: Structure) -> ElementBuilder

Create a new element-specific message. This is meant as a generic way of allowing one-way communication from an element to an application, for example "the firewire cable was unplugged". The format of the message should be documented in the element's documentation. The structure field can be None.

src

The object originating the message.

structure

The structure for the message. The message will take ownership of the structure.

Returns

The new element message.

MT safe.

fn new_segment_start(format: Format, position: i64) -> SegmentStartBuilder

Create a new segment message. This message is posted by elements that start playback of a segment as a result of a segment seek. This message is not received by the application but is used for maintenance reasons in container elements.

src

The object originating the message.

format

The format of the position being played

position

The position of the segment being played

Returns

the new segment start message.

MT safe.

fn new_segment_done(format: Format, position: i64) -> SegmentDoneBuilder

Create a new segment done message. This message is posted by elements that finish playback of a segment as a result of a segment seek. This message is received by the application after all elements that posted a segment_start have posted the segment_done.

src

The object originating the message.

format

The format of the position being done

position

The position of the segment being done

Returns

the new segment done message.

MT safe.

fn new_duration_changed() -> DurationChangedBuilder

Create a new duration changed message. This message is posted by elements that know the duration of a stream when the duration changes. This message is received by bins and is used to calculate the total duration of a pipeline.

src

The object originating the message.

Returns

The new duration-changed message.

MT safe.

fn new_latency() -> LatencyBuilder

This message can be posted by elements when their latency requirements have changed.

src

The object originating the message.

Returns

The new latency message.

MT safe.

fn new_async_start() -> AsyncStartBuilder

This message is posted by elements when they start an ASYNC state change.

src

The object originating the message.

Returns

The new async_start message.

MT safe.

fn new_async_done(running_time: u64) -> AsyncDoneBuilder

The message is posted when elements completed an ASYNC state change. running_time contains the time of the desired running_time when this elements goes to PLAYING. A value of GST_CLOCK_TIME_NONE for running_time means that the element has no clock interaction and thus doesn't care about the running_time of the pipeline.

src

The object originating the message.

running_time

the desired running_time

Returns

The new async_done message.

MT safe.

fn new_request_state(state: State) -> RequestStateBuilder

This message can be posted by elements when they want to have their state changed. A typical use case would be an audio server that wants to pause the pipeline because a higher priority stream is being played.

src

The object originating the message.

state

The new requested state

Returns

the new request state message.

MT safe.

fn new_step_start(
active: bool,
format: Format,
amount: u64,
rate: f64,
flush: bool,
intermediate: bool
) -> StepStartBuilder

This message is posted by elements when they accept or activate a new step event for amount in format.

active is set to false when the element accepted the new step event and has queued it for execution in the streaming threads.

active is set to true when the element has activated the step operation and is now ready to start executing the step in the streaming thread. After this message is emitted, the application can queue a new step operation in the element.

src

The object originating the message.

active

if the step is active or queued

format

the format of amount

amount

the amount of stepped data

rate

the rate of the stepped amount

flush

is this an flushing step

intermediate

is this an intermediate step

Returns

The new step_start message.

MT safe.

fn new_qos_builder(
live: bool,
running_time: u64,
stream_time: u64,
timestamp: u64,
duration: u64
) -> QosBuilder

fn new_progress<'a>(type_: ProgressType) -> ProgressBuilder<'a>

Progress messages are posted by elements when they use an asynchronous task to perform actions triggered by a state change.

code contains a well defined string describing the action. text should contain a user visible string detailing the current action.

src

The object originating the message.

type_

a ProgressType

code

a progress code

text

free, user visible text describing the progress

Returns

The new qos message.

fn new_toc(toc: &Toc, updated: bool) -> TocBuilder

Create a new TOC message. The message is posted by elements that discovered or updated a TOC.

src

the object originating the message.

Toc structure for the message.

updated

whether TOC was updated or not.

Returns

a new TOC message.

MT safe.

fn new_reset_time(running_time: u64) -> ResetTimeBuilder

This message is posted when the pipeline running-time should be reset to running_time, like after a flushing seek.

src

The object originating the message.

running_time

the requested running-time

Returns

The new reset_time message.

MT safe.

fn new_stream_start() -> StreamStartBuilder

Create a new stream_start message. This message is generated and posted in the sink elements of a GstBin. The bin will only forward the STREAM_START message to the application if all sinks have posted an STREAM_START message.

src

The object originating the message.

Returns

The new stream_start message.

MT safe.

fn new_need_context(context_type: &str) -> NeedContextBuilder

This message is posted when an element needs a specific Context.

src

The object originating the message.

context_type

The context type that is needed

Returns

The new need-context message.

MT safe.

fn new_have_context(context: Context) -> HaveContextBuilder

This message is posted when an element has a new local Context.

src

The object originating the message.

context

the context

Returns

The new have-context message.

MT safe.

fn new_device_added(device: &Device) -> DeviceAddedBuilder

Creates a new device-added message. The device-added message is produced by DeviceProvider or a DeviceMonitor. They announce the appearance of monitored devices.

src

The Object that created the message

device

The new Device

Returns

a newly allocated Message

fn new_device_removed(device: &Device) -> DeviceRemovedBuilder

Creates a new device-removed message. The device-removed message is produced by DeviceProvider or a DeviceMonitor. They announce the disappearance of monitored devices.

src

The Object that created the message

device

The removed Device

Returns

a newly allocated Message

fn new_property_notify<'a>(
property_name: &'a str,
value: &'a Value
) -> PropertyNotifyBuilder<'a>

Feature: v1_10

src

The Object whose property changed (may or may not be a Element)

property_name

name of the property that changed

val

new property value, or None

Returns

a newly allocated Message

fn new_stream_collection(
collection: &StreamCollection
) -> StreamCollectionBuilder

Creates a new stream-collection message. The message is used to announce new StreamCollection

Feature: v1_10

src

The Object that created the message

collection

The StreamCollection

Returns

a newly allocated Message

fn new_streams_selected(collection: &StreamCollection) -> StreamsSelectedBuilder

Creates a new steams-selected message. The message is used to announce that an array of streams has been selected. This is generally in response to a EventType::SelectStreams event, or when an element (such as decodebin3) makes an initial selection of streams.

The message also contains the StreamCollection to which the various streams belong to.

Users of Message::new_streams_selected can add the selected streams with Message::streams_selected_add.

Feature: v1_10

src

The Object that created the message

collection

The StreamCollection

Returns

a newly allocated Message

fn new_redirect<'a>(
location: &'a str,
tag_list: Option<&'a TagList>,
entry_struct: Option<Structure>
) -> RedirectBuilder<'a>

Creates a new redirect message and adds a new entry to it. Redirect messages are posted when an element detects that the actual data has to be retrieved from a different location. This is useful if such a redirection cannot be handled inside a source element, for example when HTTP 302/303 redirects return a non-HTTP URL.

The redirect message can hold multiple entries. The first one is added when the redirect message is created, with the given location, tag_list, entry_struct arguments. Use Message::add_redirect_entry to add more entries.

Each entry has a location, a tag list, and a structure. All of these are optional. The tag list and structure are useful for additional metadata, such as bitrate statistics for the given location.

By default, message recipients should treat entries in the order they are stored. The recipient should therefore try entry 0 first, and if this entry is not acceptable or working, try entry 1 etc. Senders must make sure that they add entries in this order. However, recipients are free to ignore the order and pick an entry that is "best" for them. One example would be a recipient that scans the entries for the one with the highest bitrate tag.

The specified location string is copied. However, ownership over the tag list and structure are transferred to the message.

Feature: v1_10

src

The Object whose property changed (may or may not be a Element)

location

location string for the new entry

tag_list

tag list for the new entry

entry_struct

structure for the new entry

Returns

a newly allocated Message

impl GstRc<CapsRef>

fn new_empty() -> Self

Creates a new Caps that is empty. That is, the returned Caps contains no media formats. The Caps is guaranteed to be writable. Caller is responsible for unreffing the returned caps.

Returns

the new Caps

fn new_any() -> Self

Creates a new Caps that indicates that it is compatible with any media format.

Returns

the new Caps

fn new_simple(name: &str, values: &[(&str, &ToValue)]) -> Self

Creates a new Caps that contains one Structure. The structure is defined by the arguments, which have the same format as Structure::new. Caller is responsible for unreffing the returned caps.

media_type

the media type of the structure

fieldname

first field to set

Returns

the new Caps

fn from_string(value: &str) -> Option<Self>

Converts caps from a string representation.

The current implementation of serialization will lead to unexpected results when there are nested Caps / Structure deeper than one level.

string

a string to convert to Caps

Returns

a newly allocated Caps

fn fixate(caps: Self) -> Self

fn merge(caps: Self, other: Self) -> Self

fn merge_structure(caps: Self, other: Structure) -> Self

fn normalize(caps: Self) -> Self

fn simplify(caps: Self) -> Self

fn subtract(caps: Self, other: Self) -> Self

fn truncate(caps: Self) -> Self

impl GstRc<TagListRef>

fn new() -> Self

Creates a new taglist and appends the values for the given tags. It expects tag-value pairs like TagList::add, and a None terminator after the last pair. The type of the values is implicit and is documented in the API reference, but can also be queried at runtime with gst_tag_get_type. It is an error to pass a value of a type not matching the tag type into this function. The tag list will make copies of any arguments passed (e.g. strings, buffers).

After creation you might also want to set a TagScope on the returned taglist to signal if the contained tags are global or stream tags. By default stream scope is assumes. See TagList::set_scope.

Free-function: gst_tag_list_unref

tag

tag

Returns

a new TagList. Free with gst_tag_list_unref when no longer needed.

impl GstRc<BufferRef>

fn new() -> Self

Creates a newly allocated buffer without any data.

MT safe.

Returns

the new Buffer.

fn with_size(size: usize) -> Option<Self>

fn from_vec(vec: Vec<u8>) -> Option<Self>

fn into_mapped_buffer_readable(self) -> Result<MappedBuffer<Readable>, Self>

fn into_mapped_buffer_writable(self) -> Result<MappedBuffer<Writable>, Self>

fn append(buffer: Self, other: Self) -> Self

Append all the memory from buf2 to self. The result buffer will contain a concatenation of the memory of self and buf2.

buf2

the second source Buffer to append.

Returns

the new Buffer that contains the memory of the two source buffers.

impl GstRc<SampleRef>

fn new(
buffer: Option<Buffer>,
caps: Option<Caps>,
segment: Option<&Segment>,
info: Option<&StructureRef>
) -> Self

Create a new Sample with the provided details.

Free-function: gst_sample_unref

buffer

a Buffer, or None

caps

a Caps, or None

segment

a Segment, or None

info

a Structure, or None

Returns

the new Sample. gst_sample_unref after usage.

impl GstRc<BufferListRef>

fn new() -> Self

Creates a new, empty BufferList. The caller is responsible for unreffing the returned BufferList.

Free-function: gst_buffer_list_unref

Returns

the new BufferList. gst_buffer_list_unref after usage.

fn new_sized(size: usize) -> Self

Creates a new, empty BufferList. The caller is responsible for unreffing the returned BufferList. The list will have size space preallocated so that memory reallocations can be avoided.

Free-function: gst_buffer_list_unref

size

an initial reserved size

Returns

the new BufferList. gst_buffer_list_unref after usage.

impl GstRc<QueryRef>

fn new_position(fmt: Format) -> Self

Constructs a new query stream position query object. Use gst_query_unref when done with it. A position query is used to query the current position of playback in the streams, in some format.

Free-function: gst_query_unref

format

the default Format for the new query

Returns

a new Query

fn new_duration(fmt: Format) -> Self

Constructs a new stream duration query object to query in the given format. Use gst_query_unref when done with it. A duration query will give the total length of the stream.

Free-function: gst_query_unref

format

the Format for this duration query

Returns

a new Query

fn new_latency() -> Self

Constructs a new latency query object. Use gst_query_unref when done with it. A latency query is usually performed by sinks to compensate for additional latency introduced by elements in the pipeline.

Free-function: gst_query_unref

Returns

a Query

fn new_seeking(fmt: Format) -> Self

Constructs a new query object for querying seeking properties of the stream.

Free-function: gst_query_unref

format

the default Format for the new query

Returns

a new Query

fn new_segment(fmt: Format) -> Self

Constructs a new segment query object. Use gst_query_unref when done with it. A segment query is used to discover information about the currently configured segment for playback.

Free-function: gst_query_unref

format

the Format for the new query

Returns

a new Query

fn new_convert(src_fmt: Format, value: i64, dest_fmt: Format) -> Self

Constructs a new convert query object. Use gst_query_unref when done with it. A convert query is used to ask for a conversion between one format and another.

Free-function: gst_query_unref

src_format

the source Format for the new query

value

the value to convert

dest_format

the target Format

Returns

a Query

fn new_formats() -> Self

Constructs a new query object for querying formats of the stream.

Free-function: gst_query_unref

Returns

a new Query

fn new_buffering(fmt: Format) -> Self

Constructs a new query object for querying the buffering status of a stream.

Free-function: gst_query_unref

format

the default Format for the new query

Returns

a new Query

fn new_custom(structure: Structure) -> Self

Constructs a new custom query object. Use gst_query_unref when done with it.

Free-function: gst_query_unref

type_

the query type

structure

a structure for the query

Returns

a new Query

fn new_uri() -> Self

Constructs a new query URI query object. Use gst_query_unref when done with it. An URI query is used to query the current URI that is used by the source or sink.

Free-function: gst_query_unref

Returns

a new Query

fn new_scheduling() -> Self

Constructs a new query object for querying the scheduling properties.

Free-function: gst_query_unref

Returns

a new Query

fn new_accept_caps(caps: &Caps) -> Self

Constructs a new query object for querying if caps are accepted.

Free-function: gst_query_unref

caps

a fixed Caps

Returns

a new Query

fn new_caps(filter: &Caps) -> Self

Constructs a new query object for querying the caps.

The CAPS query should return the allowable caps for a pad in the context of the element's state, its link to other elements, and the devices or files it has opened. These caps must be a subset of the pad template caps. In the NULL state with no links, the CAPS query should ideally return the same caps as the pad template. In rare circumstances, an object property can affect the caps returned by the CAPS query, but this is discouraged.

For most filters, the caps returned by CAPS query is directly affected by the allowed caps on other pads. For demuxers and decoders, the caps returned by the srcpad's getcaps function is directly related to the stream data. Again, the CAPS query should return the most specific caps it reasonably can, since this helps with autoplugging.

The filter is used to restrict the result caps, only the caps matching filter should be returned from the CAPS query. Specifying a filter might greatly reduce the amount of processing an element needs to do.

Free-function: gst_query_unref

filter

a filter

Returns

a new Query

fn new_drain() -> Self

Constructs a new query object for querying the drain state.

Free-function: gst_query_unref

Returns

a new Query

fn new_context(context_type: &str) -> Self

Constructs a new query object for querying the pipeline-local context.

Free-function: gst_query_unref

context_type

Context type to query

Returns

a new Query

impl GstRc<EventRef>

fn new_flush_start() -> FlushStartBuilder

Allocate a new flush start event. The flush start event can be sent upstream and downstream and travels out-of-bounds with the dataflow.

It marks pads as being flushing and will make them return FlowReturn::Flushing when used for data flow with Pad::push, Pad::chain, Pad::get_range and Pad::pull_range. Any event (except a EventType::FlushStop) received on a flushing pad will return false immediately.

Elements should unlock any blocking functions and exit their streaming functions as fast as possible when this event is received.

This event is typically generated after a seek to flush out all queued data in the pipeline so that the new media is played as soon as possible.

Returns

a new flush start event.

fn new_flush_stop(reset_time: bool) -> FlushStopBuilder

Allocate a new flush stop event. The flush stop event can be sent upstream and downstream and travels serialized with the dataflow. It is typically sent after sending a FLUSH_START event to make the pads accept data again.

Elements can process this event synchronized with the dataflow since the preceding FLUSH_START event stopped the dataflow.

This event is typically generated to complete a seek and to resume dataflow.

reset_time

if time should be reset

Returns

a new flush stop event.

fn new_stream_start(stream_id: &str) -> StreamStartBuilder

Create a new STREAM_START event. The stream start event can only travel downstream synchronized with the buffer flow. It is expected to be the first event that is sent for a new stream.

Source elements, demuxers and other elements that create new streams are supposed to send this event as the first event of a new stream. It should not be sent after a flushing seek or in similar situations and is used to mark the beginning of a new logical stream. Elements combining multiple streams must ensure that this event is only forwarded downstream once and not for every single input stream.

The stream_id should be a unique string that consists of the upstream stream-id, / as separator and a unique stream-id for this specific stream. A new stream-id should only be created for a stream if the upstream stream is split into (potentially) multiple new streams, e.g. in a demuxer, but not for every single element in the pipeline. PadExt::create_stream_id or PadExt::create_stream_id_printf can be used to create a stream-id. There are no particular semantics for the stream-id, though it should be deterministic (to support stream matching) and it might be used to order streams (besides any information conveyed by stream flags).

stream_id

Identifier for this stream

Returns

the new STREAM_START event.

fn new_caps(caps: &Caps) -> CapsBuilder

Create a new CAPS event for caps. The caps event can only travel downstream synchronized with the buffer flow and contains the format of the buffers that will follow after the event.

caps

a Caps

Returns

the new CAPS event.

fn new_segment(segment: &Segment) -> SegmentBuilder

Create a new SEGMENT event for segment. The segment event can only travel downstream synchronized with the buffer flow and contains timing information and playback properties for the buffers that will follow.

The segment event marks the range of buffers to be processed. All data not within the segment range is not to be processed. This can be used intelligently by plugins to apply more efficient methods of skipping unneeded data. The valid range is expressed with the start and stop values.

The time value of the segment is used in conjunction with the start value to convert the buffer timestamps into the stream time. This is usually done in sinks to report the current stream_time. time represents the stream_time of a buffer carrying a timestamp of start. time cannot be -1.

start cannot be -1, stop can be -1. If there is a valid stop given, it must be greater or equal the start, including when the indicated playback rate is < 0.

The applied_rate value provides information about any rate adjustment that has already been made to the timestamps and content on the buffers of the stream. (rate * applied_rate) should always equal the rate that has been requested for playback. For example, if an element has an input segment with intended playback rate of 2.0 and applied_rate of 1.0, it can adjust incoming timestamps and buffer content by half and output a segment event with rate of 1.0 and applied_rate of 2.0

After a segment event, the buffer stream time is calculated with:

time + (TIMESTAMP(buf) - start) * ABS (rate * applied_rate)

segment

a Segment

Returns

the new SEGMENT event.

fn new_stream_collection(
stream_collection: &StreamCollection
) -> StreamCollectionBuilder

Create a new STREAM_COLLECTION event. The stream collection event can only travel downstream synchronized with the buffer flow.

Source elements, demuxers and other elements that manage collections of streams and post StreamCollection messages on the bus also send this event downstream on each pad involved in the collection, so that activation of a new collection can be tracked through the downstream data flow.

Feature: v1_10

collection

Active collection for this data flow

Returns

the new STREAM_COLLECTION event.

fn new_tag(tags: TagList) -> TagBuilder

Generates a metadata tag event from the given taglist.

The scope of the taglist specifies if the taglist applies to the complete medium or only to this specific stream. As the tag event is a sticky event, elements should merge tags received from upstream with a given scope with their own tags with the same scope and create a new tag event from it.

taglist

metadata list. The event will take ownership of the taglist.

Returns

a new Event

fn new_buffer_size(
format: Format,
minsize: i64,
maxsize: i64,
async: bool
) -> BufferSizeBuilder

Create a new buffersize event. The event is sent downstream and notifies elements that they should provide a buffer of the specified dimensions.

When the async flag is set, a thread boundary is preferred.

format

buffer format

minsize

minimum buffer size

maxsize

maximum buffer size

async

thread behavior

Returns

a new Event

fn new_sink_message<'a>(
name: &'a str,
msg: &'a Message
) -> SinkMessageBuilder<'a>

Create a new sink-message event. The purpose of the sink-message event is to instruct a sink to post the message contained in the event synchronized with the stream.

name is used to store multiple sticky events on one pad.

name

a name for the event

msg

the Message to be posted

Returns

a new Event

fn new_stream_group_done(group_id: u32) -> StreamGroupDoneBuilder

Create a new Stream Group Done event. The stream-group-done event can only travel downstream synchronized with the buffer flow. Elements that receive the event on a pad should handle it mostly like EOS, and emit any data or pending buffers that would depend on more data arriving and unblock, since there won't be any more data.

This event is followed by EOS at some point in the future, and is generally used when switching pads - to unblock downstream so that new pads can be exposed before sending EOS on the existing pads.

Feature: v1_10

group_id

the group id of the stream group which is ending

Returns

the new stream-group-done event.

fn new_eos() -> EosBuilder

Create a new EOS event. The eos event can only travel downstream synchronized with the buffer flow. Elements that receive the EOS event on a pad can return FlowReturn::Eos as a FlowReturn when data after the EOS event arrives.

The EOS event will travel down to the sink elements in the pipeline which will then post the MessageType::Eos on the bus after they have finished playing any buffered data.

When all sinks have posted an EOS message, an EOS message is forwarded to the application.

The EOS event itself will not cause any state transitions of the pipeline.

Returns

the new EOS event.

fn new_toc(toc: &Toc, updated: bool) -> TocBuilder

Generate a TOC event from the given toc. The purpose of the TOC event is to inform elements that some kind of the TOC was found.

toc

Toc structure.

updated

whether toc was updated or not.

Returns

a new Event.

fn new_protection<'a>(
system_id: &'a str,
data: &'a Buffer,
origin: &'a str
) -> ProtectionBuilder<'a>

Creates a new event containing information specific to a particular protection system (uniquely identified by system_id), by which that protection system can acquire key(s) to decrypt a protected stream.

In order for a decryption element to decrypt media protected using a specific system, it first needs all the protection system specific information necessary to acquire the decryption key(s) for that stream. The functions defined here enable this information to be passed in events from elements that extract it (e.g., ISOBMFF demuxers, MPEG DASH demuxers) to protection decrypter elements that use it.

Events containing protection system specific information are created using Event::new_protection, and they can be parsed by downstream elements using Event::parse_protection.

In Common Encryption, protection system specific information may be located within ISOBMFF files, both in movie (moov) boxes and movie fragment (moof) boxes; it may also be contained in ContentProtection elements within MPEG DASH MPDs. The events created by Event::new_protection contain data identifying from which of these locations the encapsulated protection system specific information originated. This origin information is required as some protection systems use different encodings depending upon where the information originates.

The events returned by Event::new_protection are implemented in such a way as to ensure that the most recently-pushed protection info event of a particular origin and system_id will be stuck to the output pad of the sending element.

system_id

a string holding a UUID that uniquely identifies a protection system.

data

a Buffer holding protection system specific information. The reference count of the buffer will be incremented by one.

origin

a string indicating where the protection information carried in the event was extracted from. The allowed values of this string will depend upon the protection scheme.

Returns

a EventType::Protection event, if successful; None if unsuccessful.

fn new_segment_done(format: Format, position: i64) -> SegmentDoneBuilder

Create a new segment-done event. This event is sent by elements that finish playback of a segment as a result of a segment seek.

format

The format of the position being done

position

The position of the segment being done

Returns

a new Event

fn new_gap(timestamp: u64, duration: u64) -> GapBuilder

Create a new GAP event. A gap event can be thought of as conceptually equivalent to a buffer to signal that there is no data for a certain amount of time. This is useful to signal a gap to downstream elements which may wait for data, such as muxers or mixers or overlays, especially for sparse streams such as subtitle streams.

timestamp

the start time (pts) of the gap

duration

the duration of the gap

Returns

the new GAP event.

fn new_qos(
type_: QOSType,
proportion: f64,
diff: i64,
timestamp: u64
) -> QosBuilder

Allocate a new qos event with the given values. The QOS event is generated in an element that wants an upstream element to either reduce or increase its rate because of high/low CPU load or other resource usage such as network performance or throttling. Typically sinks generate these events for each buffer they receive.

type_ indicates the reason for the QoS event. QOSType::Overflow is used when a buffer arrived in time or when the sink cannot keep up with the upstream datarate. QOSType::Underflow is when the sink is not receiving buffers fast enough and thus has to drop late buffers. QOSType::Throttle is used when the datarate is artificially limited by the application, for example to reduce power consumption.

proportion indicates the real-time performance of the streaming in the element that generated the QoS event (usually the sink). The value is generally computed based on more long term statistics about the streams timestamps compared to the clock. A value < 1.0 indicates that the upstream element is producing data faster than real-time. A value > 1.0 indicates that the upstream element is not producing data fast enough. 1.0 is the ideal proportion value. The proportion value can safely be used to lower or increase the quality of the element.

diff is the difference against the clock in running time of the last buffer that caused the element to generate the QOS event. A negative value means that the buffer with timestamp arrived in time. A positive value indicates how late the buffer with timestamp was. When throttling is enabled, diff will be set to the requested throttling interval.

timestamp is the timestamp of the last buffer that cause the element to generate the QOS event. It is expressed in running time and thus an ever increasing value.

The upstream element can use the diff and timestamp values to decide whether to process more buffers. For positive diff, all buffers with timestamp <= timestamp + diff will certainly arrive late in the sink as well. A (negative) diff value so that timestamp + diff would yield a result smaller than 0 is not allowed.

The application can use general event probes to intercept the QoS event and implement custom application specific QoS handling.

type_

the QoS type

proportion

the proportion of the qos message

diff

The time difference of the last Clock sync

timestamp

The timestamp of the buffer

Returns

a new QOS event.

fn new_seek(
rate: f64,
format: Format,
flags: SeekFlags,
start_type: SeekType,
start: i64,
stop_type: SeekType,
stop: i64
) -> SeekBuilder

Allocate a new seek event with the given parameters.

The seek event configures playback of the pipeline between start to stop at the speed given in rate, also called a playback segment. The start and stop values are expressed in format.

A rate of 1.0 means normal playback rate, 2.0 means double speed. Negatives values means backwards playback. A value of 0.0 for the rate is not allowed and should be accomplished instead by PAUSING the pipeline.

A pipeline has a default playback segment configured with a start position of 0, a stop position of -1 and a rate of 1.0. The currently configured playback segment can be queried with QueryType::Segment.

start_type and stop_type specify how to adjust the currently configured start and stop fields in playback segment. Adjustments can be made relative or absolute to the last configured values. A type of SeekType::None means that the position should not be updated.

When the rate is positive and start has been updated, playback will start from the newly configured start position.

For negative rates, playback will start from the newly configured stop position (if any). If the stop position is updated, it must be different from -1 (GST_CLOCK_TIME_NONE) for negative rates.

It is not possible to seek relative to the current playback position, to do this, PAUSE the pipeline, query the current playback position with QueryType::Position and update the playback segment current position with a SeekType::Set to the desired position.

rate

The new playback rate

format

The format of the seek values

flags

The optional seek flags

start_type

The type and flags for the new start position

start

The value of the new start position

stop_type

The type and flags for the new stop position

stop

The value of the new stop position

Returns

a new seek event.

fn new_navigation(structure: Structure) -> NavigationBuilder

Create a new navigation event from the given description.

structure

description of the event. The event will take ownership of the structure.

Returns

a new Event

fn new_latency(latency: u64) -> LatencyBuilder

Create a new latency event. The event is sent upstream from the sinks and notifies elements that they should add an additional latency to the running time before synchronising against the clock.

The latency is mostly used in live sinks and is always expressed in the time format.

latency

the new latency value

Returns

a new Event

fn new_step(
format: Format,
amount: u64,
rate: f64,
flush: bool,
intermediate: bool
) -> StepBuilder

Create a new step event. The purpose of the step event is to instruct a sink to skip amount (expressed in format) of media. It can be used to implement stepping through the video frame by frame or for doing fast trick modes.

A rate of <= 0.0 is not allowed. Pause the pipeline, for the effect of rate = 0.0 or first reverse the direction of playback using a seek event to get the same effect as rate < 0.0.

The flush flag will clear any pending data in the pipeline before starting the step operation.

The intermediate flag instructs the pipeline that this step operation is part of a larger step operation.

format

the format of amount

amount

the amount of data to step

rate

the step rate

flush

flushing steps

intermediate

intermediate steps

Returns

a new Event

fn new_reconfigure() -> ReconfigureBuilder

Create a new reconfigure event. The purpose of the reconfigure event is to travel upstream and make elements renegotiate their caps or reconfigure their buffer pools. This is useful when changing properties on elements or changing the topology of the pipeline.

Returns

a new Event

fn new_toc_select(uid: &str) -> TocSelectBuilder

Generate a TOC select event with the given uid. The purpose of the TOC select event is to start playback based on the TOC's entry with the given uid.

uid

UID in the TOC to start playback from.

Returns

a new Event.

fn new_select_streams<'a>(streams: &'a [&'a str]) -> SelectStreamsBuilder<'a>

Allocate a new select-streams event.

The select-streams event requests the specified streams to be activated.

The list of streams corresponds to the "Stream ID" of each stream to be activated. Those ID can be obtained via the Stream objects present in EventType::StreamStart, EventType::StreamCollection or GST_MESSSAGE_STREAM_COLLECTION.

Feature: v1_10

streams

the list of streams to activate

Returns

a new select-streams event.

fn new_custom_upstream(structure: Structure) -> CustomUpstreamBuilder

fn new_custom_downstream(structure: Structure) -> CustomDownstreamBuilder

fn new_custom_downstream_oob(structure: Structure) -> CustomDownstreamOobBuilder

fn new_custom_downstream_sticky(
structure: Structure
) -> CustomDownstreamStickyBuilder

fn new_custom_both(structure: Structure) -> CustomBothBuilder

fn new_custom_both_oob(structure: Structure) -> CustomBothOobBuilder

impl GstRc<ContextRef>

fn new(context_type: &str, persistent: bool) -> Self

Create a new context.

context_type

Context type

persistent

Persistent context

Returns

The new context.

impl GstRc<TocRef>

fn new(scope: TocScope) -> Self

Create a new Toc structure.

scope

scope of this TOC

Returns

newly allocated Toc structure, free it with gst_toc_unref.

impl GstRc<TocEntryRef>

fn new(type_: TocEntryType, uid: &str) -> Self

Create new TocEntry structure.

type_

entry type.

uid

unique ID (UID) in the whole TOC.

Returns

newly allocated TocEntry structure, free it with gst_toc_entry_unref.