gstreamer::object

Trait GstObjectExtManual

pub trait GstObjectExtManual: IsA<Object> + 'static {
    // Provided methods
    fn connect_deep_notify<F: Fn(&Self, &Object, &ParamSpec) + Send + Sync + 'static>(
        &self,
        name: Option<&str>,
        f: F,
    ) -> SignalHandlerId { ... }
    fn set_object_flags(&self, flags: ObjectFlags) { ... }
    fn unset_object_flags(&self, flags: ObjectFlags) { ... }
    fn object_flags(&self) -> ObjectFlags { ... }
    fn g_value_array(
        &self,
        property_name: &str,
        timestamp: ClockTime,
        interval: ClockTime,
        values: &mut [Value],
    ) -> Result<(), BoolError> { ... }
    fn object_lock(&self) -> ObjectLockGuard<'_, Self> { ... }
    unsafe fn set_parent(
        &self,
        parent: &impl IsA<Object>,
    ) -> Result<Self, BoolError> { ... }
    unsafe fn unset_parent(
        &self,
        parent: &impl IsA<Object>,
    ) -> Result<(), BoolError> { ... }
}

Provided Methods

fn connect_deep_notify<F: Fn(&Self, &Object, &ParamSpec) + Send + Sync + 'static>( &self, name: Option<&str>, f: F, ) -> SignalHandlerId

fn set_object_flags(&self, flags: ObjectFlags)

fn unset_object_flags(&self, flags: ObjectFlags)

fn object_flags(&self) -> ObjectFlags

fn g_value_array( &self, property_name: &str, timestamp: ClockTime, interval: ClockTime, values: &mut [Value], ) -> Result<(), BoolError>

fn object_lock(&self) -> ObjectLockGuard<'_, Self>

unsafe fn set_parent( &self, parent: &impl IsA<Object>, ) -> Result<Self, BoolError>

Sets the parent of self to parent. If self already has a parent this will fail.

The returned reference to self on success must be kept alive until the parent is unset again. Also parent must be kept alive until child gets its parent unset, or in other words if child doesn’t get its parent unset until parent is disposed then during disposal this must happen. Not doing so causes dangling references and potential use-after-frees.

unsafe fn unset_parent( &self, parent: &impl IsA<Object>, ) -> Result<(), BoolError>

Unsets the parent of self from parent. If self has no parent or a different parent than parent this will fail.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.