Trait gstreamer::prelude::ObjectExt

pub trait ObjectExt: ObjectType {
    fn is<T>(&self) -> bool
    where
        T: StaticType;

    fn get_type(&self) -> Type;

    fn get_object_class(&self) -> &ObjectClass;

    fn set_property<'a, N>(
        &self, 
        property_name: N, 
        value: &dyn ToValue
    ) -> Result<(), BoolError>
    where
        N: Into<&'a str>;

    fn set_properties(
        &self, 
        property_values: &[(&str, &dyn ToValue)]
    ) -> Result<(), BoolError>;

    fn get_property<'a, N>(&self, property_name: N) -> Result<Value, BoolError>
    where
        N: Into<&'a str>;

    fn has_property<'a, N>(&self, property_name: N, type_: Option<Type>) -> bool
    where
        N: Into<&'a str>;

    fn get_property_type<'a, N>(&self, property_name: N) -> Option<Type>
    where
        N: Into<&'a str>;

    fn find_property<'a, N>(&self, property_name: N) -> Option<ParamSpec>
    where
        N: Into<&'a str>;

    fn list_properties(&self) -> Vec<ParamSpec>;

    unsafe fn set_qdata<QD>(&self, key: Quark, value: QD)
    where
        QD: 'static;

    unsafe fn get_qdata<QD>(&self, key: Quark) -> Option<&QD>
    where
        QD: 'static;

    unsafe fn steal_qdata<QD>(&self, key: Quark) -> Option<QD>
    where
        QD: 'static;

    unsafe fn set_data<QD>(&self, key: &str, value: QD)
    where
        QD: 'static;

    unsafe fn get_data<QD>(&self, key: &str) -> Option<&QD>
    where
        QD: 'static;

    unsafe fn steal_data<QD>(&self, key: &str) -> Option<QD>
    where
        QD: 'static;

    fn block_signal(&self, handler_id: &SignalHandlerId);

    fn unblock_signal(&self, handler_id: &SignalHandlerId);

    fn stop_signal_emission(&self, signal_name: &str);

    fn connect<'a, N, F>(
        &self, 
        signal_name: N, 
        after: bool, 
        callback: F
    ) -> Result<SignalHandlerId, BoolError>
    where
        F: Fn(&[Value]) -> Option<Value> + Send + Sync + 'static,
        N: Into<&'a str>;

    fn connect_local<'a, N, F>(
        &self, 
        signal_name: N, 
        after: bool, 
        callback: F
    ) -> Result<SignalHandlerId, BoolError>
    where
        F: Fn(&[Value]) -> Option<Value> + 'static,
        N: Into<&'a str>;

    unsafe fn connect_unsafe<'a, N, F>(
        &self, 
        signal_name: N, 
        after: bool, 
        callback: F
    ) -> Result<SignalHandlerId, BoolError>
    where
        F: Fn(&[Value]) -> Option<Value>,
        N: Into<&'a str>;

    fn emit<'a, N>(
        &self, 
        signal_name: N, 
        args: &[&dyn ToValue]
    ) -> Result<Option<Value>, BoolError>
    where
        N: Into<&'a str>;

    fn disconnect(&self, handler_id: SignalHandlerId);

    fn connect_notify<F>(&self, name: Option<&str>, f: F) -> SignalHandlerId
    where
        F: 'static + Send + Sync + Fn(&Self, &ParamSpec);

    unsafe fn connect_notify_unsafe<F>(
        &self, 
        name: Option<&str>, 
        f: F
    ) -> SignalHandlerId
    where
        F: Fn(&Self, &ParamSpec);

    fn notify<'a, N>(&self, property_name: N)
    where
        N: Into<&'a str>;

    fn notify_by_pspec(&self, pspec: &ParamSpec);

    fn downgrade(&self) -> WeakRef<Self>;

    fn bind_property<'a, O, N, M>(
        &'a self, 
        source_property: N, 
        target: &'a O, 
        target_property: M
    ) -> BindingBuilder<'a>
    where
        M: Into<&'a str>,
        N: Into<&'a str>,
        O: ObjectType;

    fn ref_count(&self) -> u32;
}

Required methods

fn is<T>(&self) -> bool where
T: StaticType,

Returns true if the object is an instance of (can be cast to) T.

fn get_type(&self) -> Type

fn get_object_class(&self) -> &ObjectClass

fn set_property<'a, N>(
&self,
property_name: N,
value: &dyn ToValue
) -> Result<(), BoolError> where
N: Into<&'a str>,

fn set_properties(
&self,
property_values: &[(&str, &dyn ToValue)]
) -> Result<(), BoolError>

fn get_property<'a, N>(&self, property_name: N) -> Result<Value, BoolError> where
N: Into<&'a str>,

fn has_property<'a, N>(&self, property_name: N, type_: Option<Type>) -> bool where
N: Into<&'a str>,

fn get_property_type<'a, N>(&self, property_name: N) -> Option<Type> where
N: Into<&'a str>,

fn find_property<'a, N>(&self, property_name: N) -> Option<ParamSpec> where
N: Into<&'a str>,

fn list_properties(&self) -> Vec<ParamSpec>

unsafe fn set_qdata<QD>(&self, key: Quark, value: QD) where
QD: 'static,

Safety

This function doesn't store type information

unsafe fn get_qdata<QD>(&self, key: Quark) -> Option<&QD> where
QD: 'static,

Safety

The caller is responsible for ensuring the returned value is of a suitable type

unsafe fn steal_qdata<QD>(&self, key: Quark) -> Option<QD> where
QD: 'static,

Safety

The caller is responsible for ensuring the returned value is of a suitable type

unsafe fn set_data<QD>(&self, key: &str, value: QD) where
QD: 'static,

Safety

This function doesn't store type information

unsafe fn get_data<QD>(&self, key: &str) -> Option<&QD> where
QD: 'static,

Safety

The caller is responsible for ensuring the returned value is of a suitable type

unsafe fn steal_data<QD>(&self, key: &str) -> Option<QD> where
QD: 'static,

Safety

The caller is responsible for ensuring the returned value is of a suitable type

fn block_signal(&self, handler_id: &SignalHandlerId)

fn unblock_signal(&self, handler_id: &SignalHandlerId)

fn stop_signal_emission(&self, signal_name: &str)

fn connect<'a, N, F>(
&self,
signal_name: N,
after: bool,
callback: F
) -> Result<SignalHandlerId, BoolError> where
F: Fn(&[Value]) -> Option<Value> + Send + Sync + 'static,
N: Into<&'a str>,

fn connect_local<'a, N, F>(
&self,
signal_name: N,
after: bool,
callback: F
) -> Result<SignalHandlerId, BoolError> where
F: Fn(&[Value]) -> Option<Value> + 'static,
N: Into<&'a str>,

unsafe fn connect_unsafe<'a, N, F>(
&self,
signal_name: N,
after: bool,
callback: F
) -> Result<SignalHandlerId, BoolError> where
F: Fn(&[Value]) -> Option<Value>,
N: Into<&'a str>,

fn emit<'a, N>(
&self,
signal_name: N,
args: &[&dyn ToValue]
) -> Result<Option<Value>, BoolError> where
N: Into<&'a str>,

fn disconnect(&self, handler_id: SignalHandlerId)

fn connect_notify<F>(&self, name: Option<&str>, f: F) -> SignalHandlerId where
F: 'static + Send + Sync + Fn(&Self, &ParamSpec),

unsafe fn connect_notify_unsafe<F>(
&self,
name: Option<&str>,
f: F
) -> SignalHandlerId where
F: Fn(&Self, &ParamSpec),

fn notify<'a, N>(&self, property_name: N) where
N: Into<&'a str>,

fn notify_by_pspec(&self, pspec: &ParamSpec)

fn downgrade(&self) -> WeakRef<Self>

fn bind_property<'a, O, N, M>(
&'a self,
source_property: N,
target: &'a O,
target_property: M
) -> BindingBuilder<'a> where
M: Into<&'a str>,
N: Into<&'a str>,
O: ObjectType,

fn ref_count(&self) -> u32

Loading content...