Trait gstreamer::prelude::OptionOverflowingSub

pub trait OptionOverflowingSub<Rhs = Self, InnerRhs = Rhs> {
    type Output;

    // Required method
    fn opt_overflowing_sub(self, rhs: Rhs) -> Option<(Self::Output, bool)>;
}

Trait for values and Options overflowing substraction.

Implementing this trait leads to the following auto-implementations:

OptionOverflowingSub<Option<InnerRhs>> for T.

OptionOverflowingSub<Rhs> for Option<T>.

OptionOverflowingSub<Option<InnerRhs>> for Option<T>.

… and some variants with references.

Note that since the std library doesn’t define any OverflowingSub trait, users must provide the base implementation for the inner type.

Required Associated Types

type Output

The resulting inner type after applying the substraction.

Required Methods

fn opt_overflowing_sub(self, rhs: Rhs) -> Option<(Self::Output, bool)>

Returns a tuple of the substraction along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then self is returned.

Returns None if at least one argument is None.