Module gstreamer::query

Structs

AcceptCaps

Allocation

Buffering

Caps

Context
Context is a container object used to store contexts like a device context, a display server connection and similar concepts that should be shared between multiple elements.

Convert

Custom

Drain

Duration

Formats

Jitter

Latency

Other

Position

QueryRef

Rate

Scheduling

Seeking

Segment

Uri

Enums

QueryView

Type Definitions

Query