Struct gstreamer::DeviceProviderFactory

pub struct DeviceProviderFactory(_, _);

DeviceProviderFactory is used to create instances of device providers. A GstDeviceProviderfactory can be added to a Plugin as it is also a PluginFeature.

Use the DeviceProviderFactory::find and DeviceProviderFactory::get functions to create device provider instances or use DeviceProviderFactory::get_by_name as a convenient shortcut.

Implements

ObjectExt, ObjectExt

Methods

impl DeviceProviderFactory

fn get(&self) -> Option<DeviceProvider>

Returns the device provider of the type defined by the given device providerfactory.

Returns

the DeviceProvider or None if the device provider couldn't be created

fn get_device_provider_type(&self) -> Type

Get the glib::Type for device providers managed by this factory. The type can only be retrieved if the device provider factory is loaded, which can be assured with PluginFeature::load.

Returns

the glib::Type for device providers managed by this factory.

fn get_metadata(&self, key: &str) -> Option<String>

Get the metadata on self with key.

key

a key

Returns

the metadata with key on self or None when there was no metadata with the given key.

fn get_metadata_keys(&self) -> Vec<String>

Get the available keys for the metadata on self.

Returns

a None-terminated array of key strings, or None when there is no metadata. Free with g_strfreev when no longer needed.

fn has_classes<'a, P: Into<Option<&'a str>>>(&self, classes: P) -> bool

Check if self matches all of the given classes

classes

a "/" separate list of classes to match, only match if all classes are matched

Returns

true if self matches or if classes is None.

fn has_classesv(&self, classes: &[&str]) -> bool

Check if self matches all of the given classes

classes

a None terminated array of classes to match, only match if all classes are matched

Returns

true if self matches.

fn find(name: &str) -> Option<DeviceProviderFactory>

Search for an device provider factory of the given name. Refs the returned device provider factory; caller is responsible for unreffing.

name

name of factory to find

Returns

DeviceProviderFactory if found, None otherwise

fn get_by_name(factoryname: &str) -> Option<DeviceProvider>

Returns the device provider of the type defined by the given device provider factory.

factoryname

a named factory to instantiate

Returns

a DeviceProvider or None if unable to create device provider

fn list_get_device_providers(minrank: Rank) -> Vec<DeviceProviderFactory>

Get a list of factories with a rank greater or equal to minrank. The list of factories is returned by decreasing rank.

minrank

Minimum rank

Returns

a glib::List of DeviceProviderFactory device providers. Use PluginFeature::list_free after usage.