Struct gstreamer::GhostPad

pub struct GhostPad(_, _);

GhostPads are useful when organizing pipelines with Bin like elements. The idea here is to create hierarchical element graphs. The bin element contains a sub-graph. Now one would like to treat the bin-element like any other Element. This is where GhostPads come into play. A GhostPad acts as a proxy for another pad. Thus the bin can have sink and source ghost-pads that are associated with sink and source pads of the child elements.

If the target pad is known at creation time, GhostPad::new is the function to use to get a ghost-pad. Otherwise one can use GhostPad::new_no_target to create the ghost-pad and use GhostPadExt::set_target to establish the association later on.

Note that GhostPads add overhead to the data processing of a pipeline.

Implements

GhostPadExt, ProxyPadExt, PadExt, ObjectExt, ObjectExt

Methods

impl GhostPad

fn new_no_target<'a, P: Into<Option<&'a str>>>(
name: P,
dir: PadDirection
) -> GhostPad

Create a new ghostpad without a target with the given direction. A target can be set on the ghostpad later with the GhostPadExt::set_target function.

The created ghostpad will not have a padtemplate.

name

the name of the new pad, or None to assign a default name.

dir

the direction of the ghostpad

Returns

a new Pad, or None in case of an error.

fn new_no_target_from_template<'a, P: Into<Option<&'a str>>>(
name: P,
templ: &PadTemplate
) -> GhostPad

Create a new ghostpad based on templ, without setting a target. The direction will be taken from the templ.

name

the name of the new pad, or None to assign a default name

templ

the PadTemplate to create the ghostpad from.

Returns

a new Pad, or None in case of an error.

fn activate_mode_default<'a, P: IsA<Pad>, Q: IsA<Object> + 'a, R: Into<Option<&'a Q>>>(
pad: &P,
parent: R,
mode: PadMode,
active: bool
) -> bool

Invoke the default activate mode function of a ghost pad.

pad

the Pad to activate or deactivate.

parent

the parent of pad or None

mode

the requested activation mode

active

whether the pad should be active or not.

Returns

true if the operation was successful.

fn internal_activate_mode_default<'a, P: IsA<Pad>, Q: IsA<Object> + 'a, R: Into<Option<&'a Q>>>(
pad: &P,
parent: R,
mode: PadMode,
active: bool
) -> bool

Invoke the default activate mode function of a proxy pad that is owned by a ghost pad.

pad

the Pad to activate or deactivate.

parent

the parent of pad or None

mode

the requested activation mode

active

whether the pad should be active or not.

Returns

true if the operation was successful.

impl GhostPad

fn new<'a, P: Into<Option<&'a str>>, Q: IsA<Pad>>(
name: P,
target: &Q
) -> Option<GhostPad>

Create a new ghostpad with target as the target. The direction will be taken from the target pad. target must be unlinked.

Will ref the target.

name

the name of the new pad, or None to assign a default name

target

the pad to ghost.

Returns

a new Pad, or None in case of an error.

fn new_from_template<'a, P: Into<Option<&'a str>>, Q: IsA<Pad>>(
name: P,
target: &Q,
templ: &PadTemplate
) -> Option<GhostPad>

Create a new ghostpad with target as the target. The direction will be taken from the target pad. The template used on the ghostpad will be template.

Will ref the target.

name

the name of the new pad, or None to assign a default name.

target

the pad to ghost.

templ

the PadTemplate to use on the ghostpad.

Returns

a new Pad, or None in case of an error.