Struct gstreamer::PadTemplate

pub struct PadTemplate(_, _);

Padtemplates describe the possible media types a pad or an elementfactory can handle. This allows for both inspection of handled types before loading the element plugin as well as identifying pads on elements that are not yet created (request or sometimes pads).

Pad and PadTemplates have Caps attached to it to describe the media type they are capable of dealing with. PadTemplate::get_caps or GST_PAD_TEMPLATE_CAPS() are used to get the caps of a padtemplate. It's not possible to modify the caps of a padtemplate after creation.

PadTemplates have a PadPresence property which identifies the lifetime of the pad and that can be retrieved with GST_PAD_TEMPLATE_PRESENCE(). Also the direction of the pad can be retrieved from the PadTemplate with GST_PAD_TEMPLATE_DIRECTION().

The GST_PAD_TEMPLATE_NAME_TEMPLATE () is important for GST_PAD_REQUEST pads because it has to be used as the name in the ElementExt::get_request_pad call to instantiate a pad from this template.

Padtemplates can be created with PadTemplate::new or with gst_static_pad_template_get (), which creates a PadTemplate from a StaticPadTemplate that can be filled with the convenient GST_STATIC_PAD_TEMPLATE() macro.

A padtemplate can be used to create a pad (see Pad::new_from_template or gst_pad_new_from_static_template ()) or to add to an element class (see gst_element_class_add_static_pad_template ()).

The following code example shows the code to create a pad from a padtemplate.

  GstStaticPadTemplate my_template =
  GST_STATIC_PAD_TEMPLATE (
    "sink",          // the name of the pad
    GST_PAD_SINK,    // the direction of the pad
    GST_PAD_ALWAYS,  // when this pad will be present
    GST_STATIC_CAPS (        // the capabilities of the padtemplate
      "audio/x-raw, "
        "channels = (int) [ 1, 6 ]"
    )
  );
  void
  my_method (void)
  {
    GstPad *pad;
    pad = gst_pad_new_from_static_template (&amp;my_template, "sink");
    ...
  }

The following example shows you how to add the padtemplate to an element class, this is usually done in the class_init of the class:

  static void
  my_element_class_init (GstMyElementClass *klass)
  {
    GstElementClass *gstelement_class = GST_ELEMENT_CLASS (klass);

    gst_element_class_add_static_pad_template (gstelement_class, &amp;my_template);
  }

Implements

GstObjectExt, glib::object::ObjectExt

Methods

impl PadTemplate

pub fn new(
name_template: &str,
direction: PadDirection,
presence: PadPresence,
caps: &Caps
) -> PadTemplate

Creates a new pad template with a name according to the given template and with the given arguments.

name_template

the name template.

direction

the PadDirection of the template.

presence

the PadPresence of the pad.

caps

a Caps set for the template.

Returns

a new PadTemplate.

pub fn new_with_gtype(
name_template: &str,
direction: PadDirection,
presence: PadPresence,
caps: &Caps,
pad_type: Type
) -> PadTemplate

Creates a new pad template with a name according to the given template and with the given arguments.

Feature: v1_14

name_template

the name template.

direction

the PadDirection of the template.

presence

the PadPresence of the pad.

caps

a Caps set for the template.

pad_type

The glib::Type of the pad to create

Returns

a new PadTemplate.

pub fn get_caps(&self) -> Option<Caps>

Gets the capabilities of the pad template.

Returns

the Caps of the pad template. Unref after usage.

pub fn pad_created<P: IsA<Pad>>(&self, pad: &P)

Emit the pad-created signal for this template when created by this pad.

pad

the Pad that created it

pub fn get_property_direction(&self) -> PadDirection

pub fn get_property_gtype(&self) -> Type

pub fn get_property_name_template(&self) -> Option<String>

pub fn get_property_presence(&self) -> PadPresence

pub fn connect_pad_created<F: Fn(&PadTemplate, &Pad) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_caps_notify<F: Fn(&PadTemplate) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_direction_notify<F: Fn(&PadTemplate) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_gtype_notify<F: Fn(&PadTemplate) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_name_template_notify<F: Fn(&PadTemplate) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_presence_notify<F: Fn(&PadTemplate) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

impl PadTemplate

pub fn new_from_static_pad_template_with_gtype(
pad_template: &StaticPadTemplate,
pad_type: Type
) -> PadTemplate

Converts a StaticPadTemplate into a PadTemplate with a type.

Feature: v1_14

pad_template

the static pad template

pad_type

The glib::Type of the pad to create

Returns

a new PadTemplate.