Struct gstreamer::Plugin

pub struct Plugin(_, _);

GStreamer is extensible, so Element instances can be loaded at runtime. A plugin system can provide one or more of the basic <application>GStreamer</application> PluginFeature subclasses.

A plugin should export a symbol <symbol>gst_plugin_desc</symbol> that is a struct of type PluginDesc. the plugin loader will check the version of the core library the plugin was linked against and will create a new Plugin. It will then call the GstPluginInitFunc function that was provided in the <symbol>gst_plugin_desc</symbol>.

Once you have a handle to a Plugin (e.g. from the Registry), you can add any object that subclasses PluginFeature.

Usually plugins are always automatically loaded so you don't need to call Plugin::load explicitly to bring it into memory. There are options to statically link plugins to an app or even use GStreamer without a plugin repository in which case Plugin::load can be needed to bring the plugin into memory.

Implements

ObjectExt, ObjectExt

Methods

impl Plugin

pub fn get_cache_data(&self) -> Option<Structure>

Gets the plugin specific data cache. If it is None there is no cached data stored. This is the case when the registry is getting rebuilt.

Returns

The cached data as a Structure or None.

pub fn get_description(&self) -> Option<String>

Get the long descriptive name of the plugin

Returns

the long name of the plugin

pub fn get_filename(&self) -> Option<String>

get the filename of the plugin

Returns

the filename of the plugin

pub fn get_license(&self) -> Option<String>

get the license of the plugin

Returns

the license of the plugin

pub fn get_origin(&self) -> Option<String>

get the URL where the plugin comes from

Returns

the origin of the plugin

pub fn get_package(&self) -> Option<String>

get the package the plugin belongs to.

Returns

the package of the plugin

pub fn get_release_date_string(&self) -> Option<String>

Get the release date (and possibly time) in form of a string, if available.

For normal GStreamer plugin releases this will usually just be a date in the form of "YYYY-MM-DD", while pre-releases and builds from git may contain a time component after the date as well, in which case the string will be formatted like "YYYY-MM-DDTHH:MMZ" (e.g. "2012-04-30T09:30Z").

There may be plugins that do not have a valid release date set on them.

Returns

the date string of the plugin, or None if not available.

pub fn get_source(&self) -> Option<String>

get the source module the plugin belongs to.

Returns

the source of the plugin

pub fn get_version(&self) -> Option<String>

get the version of the plugin

Returns

the version of the plugin

pub fn is_loaded(&self) -> bool

queries if the plugin is loaded into memory

Returns

true is loaded, false otherwise

pub fn load(&self) -> Option<Plugin>

Loads self. Note that the return value is the loaded plugin; self is untouched. The normal use pattern of this function goes like this:

GstPlugin *loaded_plugin;
loaded_plugin = gst_plugin_load (plugin);
// presumably, we're no longer interested in the potentially-unloaded plugin
gst_object_unref (plugin);
plugin = loaded_plugin;

Returns

a reference to a loaded plugin, or None on error.

pub fn set_cache_data(&self, cache_data: &mut Structure)

Adds plugin specific data to cache. Passes the ownership of the structure to the self.

The cache is flushed every time the registry is rebuilt.

cache_data

a structure containing the data to cache

pub fn list_free(list: &[Plugin])

Unrefs each member of list, then frees the list.

list

list of Plugin

pub fn load_by_name(name: &str) -> Option<Plugin>

Load the named plugin. Refs the plugin.

name

name of plugin to load

Returns

a reference to a loaded plugin, or None on error.

pub fn load_file(filename: &str) -> Result<Plugin, Error>

Loads the given plugin and refs it. Caller needs to unref after use.

filename

the plugin filename to load

Returns

a reference to the existing loaded GstPlugin, a reference to the newly-loaded GstPlugin, or None if an error occurred.