Struct gstreamer::ProxyPad

pub struct ProxyPad(_, _);

Implements

ProxyPadExt, PadExt, GstObjectExt, glib::object::ObjectExt

Methods

impl ProxyPad

pub fn chain_default<'a, P: IsA<ProxyPad>, Q: IsA<Object> + 'a, R: Into<Option<&'a Q>>>(
pad: &P,
parent: R,
buffer: Buffer
) -> FlowReturn

Invoke the default chain function of the proxy pad.

pad

a sink Pad, returns GST_FLOW_ERROR if not.

parent

the parent of pad or None

buffer

the Buffer to send, return GST_FLOW_ERROR if not.

Returns

a FlowReturn from the pad.

pub fn chain_list_default<'a, P: IsA<ProxyPad>, Q: IsA<Object> + 'a, R: Into<Option<&'a Q>>>(
pad: &P,
parent: R,
list: BufferList
) -> FlowReturn

Invoke the default chain list function of the proxy pad.

pad

a sink Pad, returns GST_FLOW_ERROR if not.

parent

the parent of pad or None

list

the BufferList to send, return GST_FLOW_ERROR if not.

Returns

a FlowReturn from the pad.

pub fn getrange_default<P: IsA<ProxyPad>, Q: IsA<Object>>(
pad: &P,
parent: &Q,
offset: u64,
size: u32
) -> Result<Buffer, FlowReturn>

Invoke the default getrange function of the proxy pad.

pad

a src Pad, returns FlowReturn::Error if not.

parent

the parent of pad

offset

The start offset of the buffer

size

The length of the buffer

buffer

a pointer to hold the Buffer, returns FlowReturn::Error if None.

Returns

a FlowReturn from the pad.

pub fn iterate_internal_links_default<'a, P: IsA<ProxyPad>, Q: IsA<Object> + 'a, R: Into<Option<&'a Q>>>(
pad: &P,
parent: R
) -> Option<Iterator<Pad>>