Struct gstreamer::Segment

pub struct Segment(_);

Methods

impl Segment

pub fn new() -> Segment

Allocate a new Segment structure and initialize it using Segment::init.

Free-function: gst_segment_free

Returns

a new Segment, free with Segment::free.

pub fn clip(&self, format: Format, start: u64, stop: u64) -> Option<(u64, u64)>

Clip the given start and stop values to the segment boundaries given in self. start and stop are compared and clipped to self start and stop values.

If the function returns false, start and stop are known to fall outside of self and clip_start and clip_stop are not updated.

When the function returns true, clip_start and clip_stop will be updated. If clip_start or clip_stop are different from start or stop respectively, the region fell partially in the segment.

Note that when stop is -1, clip_stop will be set to the end of the segment. Depending on the use case, this may or may not be what you want.

format

the format of the segment.

start

the start position in the segment

stop

the stop position in the segment

clip_start

the clipped start position in the segment

clip_stop

the clipped stop position in the segment

Returns

true if the given start and stop times fall partially or completely in self, false if the values are completely outside of the segment.

pub fn copy_into(&self, dest: &mut Segment)

Copy the contents of self into dest.

dest

a Segment

pub fn do_seek(
&mut self,
rate: f64,
format: Format,
flags: SeekFlags,
start_type: SeekType,
start: u64,
stop_type: SeekType,
stop: u64
) -> Option<bool>

Update the segment structure with the field values of a seek event (see Event::new_seek).

After calling this method, the segment field position and time will contain the requested new position in the segment. The new requested position in the segment depends on rate and start_type and stop_type.

For positive rate, the new position in the segment is the new self start field when it was updated with a start_type different from SeekType::None. If no update was performed on self start position (SeekType::None), start is ignored and self position is unmodified.

For negative rate, the new position in the segment is the new self stop field when it was updated with a stop_type different from SeekType::None. If no stop was previously configured in the segment, the duration of the segment will be used to update the stop position. If no update was performed on self stop position (SeekType::None), stop is ignored and self position is unmodified.

The applied rate of the segment will be set to 1.0 by default. If the caller can apply a rate change, it should update self rate and applied_rate after calling this function.

update will be set to true if a seek should be performed to the segment position field. This field can be false if, for example, only the rate has been changed but not the playback position.

rate

the rate of the segment.

format

the format of the segment.

flags

the segment flags for the segment

start_type

the seek method

start

the seek start value

stop_type

the seek method

stop

the seek stop value

update

boolean holding whether position was updated.

Returns

true if the seek could be performed.

pub fn init(&mut self, format: Format)

pub fn offset_running_time(&mut self, format: Format, offset: i64) -> bool

pub fn position_from_running_time(
&self,
format: Format,
running_time: u64
) -> u64

pub fn position_from_running_time_full(
&self,
format: Format,
running_time: u64
) -> (i32, u64)

pub fn position_from_stream_time(&self, format: Format, stream_time: u64) -> u64

pub fn position_from_stream_time_full(
&self,
format: Format,
stream_time: u64
) -> (i32, u64)

pub fn set_running_time(&mut self, format: Format, running_time: u64) -> bool

pub fn to_position(&self, format: Format, running_time: u64) -> u64

pub fn to_running_time(&self, format: Format, position: u64) -> u64

pub fn to_running_time_full(&self, format: Format, position: u64) -> (i32, u64)

pub fn to_stream_time(&self, format: Format, position: u64) -> u64

pub fn to_stream_time_full(&self, format: Format, position: u64) -> (i32, u64)

pub fn get_flags(&self) -> SegmentFlags

pub fn set_flags(&mut self, flags: SegmentFlags)

pub fn get_rate(&self) -> f64

pub fn set_rate(&mut self, rate: f64)

pub fn get_applied_rate(&self) -> f64

pub fn set_applied_rate(&mut self, applied_rate: f64)

pub fn get_format(&self) -> Format

pub fn set_format(&mut self, format: Format)

pub fn get_base(&self) -> u64

pub fn set_base(&mut self, base: u64)

pub fn get_offset(&self) -> u64

pub fn set_offset(&mut self, offset: u64)

pub fn get_start(&self) -> u64

pub fn set_start(&mut self, start: u64)

pub fn get_stop(&self) -> u64

pub fn set_stop(&mut self, stop: u64)

pub fn get_time(&self) -> u64

pub fn set_time(&mut self, time: u64)

pub fn get_position(&self) -> u64

pub fn set_position(&mut self, position: u64)

pub fn get_duration(&self) -> u64

pub fn set_duration(&mut self, duration: u64)