Struct gstreamer::Stream

pub struct Stream(_, _);

A high-level object representing a single stream. It might be backed, or not, by an actual flow of data in a pipeline (Pad).

A Stream does not care about data changes (such as decoding, encoding, parsing,...) as long as the underlying data flow corresponds to the same high-level flow (ex: a certain audio track).

A Stream contains all the information pertinent to a stream, such as stream-id, tags, caps, type, ...

Elements can subclass a Stream for internal usage (to contain information pertinent to streams of data).

Feature: v1_10

Implements

GstObjectExt, glib::object::ObjectExt

Methods

impl Stream

pub fn get_caps(&self) -> Option<Caps>

Retrieve the caps for self, if any

Feature: v1_10

Returns

The Caps for self

pub fn get_stream_flags(&self) -> StreamFlags

Retrieve the current stream flags for self

Feature: v1_10

Returns

The StreamFlags for self

pub fn get_stream_id(&self) -> Option<String>

Returns the stream ID of self.

Feature: v1_10

Returns

the stream ID of self. Only valid during the lifetime of self.

pub fn get_stream_type(&self) -> StreamType

Retrieve the stream type for self

Feature: v1_10

Returns

The StreamType for self

pub fn get_tags(&self) -> Option<TagList>

Retrieve the tags for self, if any

Feature: v1_10

Returns

The TagList for self

pub fn set_caps<'a, P: Into<Option<&'a Caps>>>(&self, caps: P)

Set the caps for the Stream

Feature: v1_10

caps

a Caps

pub fn set_stream_flags(&self, flags: StreamFlags)

Set the flags for the self.

Feature: v1_10

flags

the flags to set on self

pub fn set_stream_type(&self, stream_type: StreamType)

Set the stream type of self

Feature: v1_10

stream_type

the type to set on self

pub fn set_tags<'a, P: Into<Option<&'a TagList>>>(&self, tags: P)

Set the tags for the Stream

Feature: v1_10

tags

a TagList

pub fn get_property_caps(&self) -> Option<Caps>

pub fn set_property_caps(&self, caps: Option<&Caps>)

pub fn get_property_stream_flags(&self) -> StreamFlags

pub fn set_property_stream_flags(&self, stream_flags: StreamFlags)

pub fn get_property_stream_id(&self) -> Option<String>

pub fn get_property_stream_type(&self) -> StreamType

pub fn set_property_stream_type(&self, stream_type: StreamType)

pub fn get_property_tags(&self) -> Option<TagList>

pub fn set_property_tags(&self, tags: Option<&TagList>)

pub fn connect_property_caps_notify<F: Fn(&Stream) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_stream_flags_notify<F: Fn(&Stream) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_stream_id_notify<F: Fn(&Stream) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_stream_type_notify<F: Fn(&Stream) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_tags_notify<F: Fn(&Stream) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

impl Stream

pub fn new<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b Caps>>>(
stream_id: P,
caps: Q,
type_: StreamType,
flags: StreamFlags
) -> Stream

Create a new Stream for the given stream_id, caps, type_ and flags

Feature: v1_10

stream_id

the id for the new stream. If None, a new one will be automatically generated

caps

the Caps of the stream

type_

the StreamType of the stream

flags

the StreamFlags of the stream

Returns

The new Stream