Struct gstreamer::StreamCollection

pub struct StreamCollection(_, _);

A collection of Stream that are available.

A StreamCollection will be provided by elements that can make those streams available. Applications can use the collection to show the user what streams are available by using StreamCollection::get_stream()

Once posted, a StreamCollection is immutable. Updates are made by sending a new StreamCollection message, which may or may not share some of the Stream objects from the collection it replaces. The receiver can check the sender of a stream collection message to know which collection is obsoleted.

Several elements in a pipeline can provide StreamCollection.

Applications can activate streams from a collection by using the EventType::SelectStreams event on a pipeline, bin or element.

Feature: v1_10

Implements

ObjectExt, ObjectExt

Methods

impl StreamCollection

fn new<'a, P: Into<Option<&'a str>>>(upstream_id: P) -> StreamCollection

Create a new StreamCollection.

Feature: v1_10

upstream_id

The stream id of the parent stream

Returns

The new StreamCollection.

fn add_stream(&self, stream: &Stream) -> bool

Add the given stream to the self.

Feature: v1_10

stream

the Stream to add

Returns

true if the stream was properly added, else false

fn get_size(&self) -> u32

Get the number of streams this collection contains

Feature: v1_10

Returns

The number of streams that self contains

fn get_stream(&self, index: u32) -> Option<Stream>

Retrieve the Stream with index index from the collection.

The caller should not modify the returned Stream

Feature: v1_10

index

Index of the stream to retrieve

Returns

A Stream

fn get_upstream_id(&self) -> Option<String>

Returns the upstream id of the self.

Feature: v1_10

Returns

The upstream id

fn get_property_upstream_id(&self) -> Option<String>

fn set_property_upstream_id(&self, upstream_id: Option<&str>)

fn connect_property_upstream_id_notify<F: Fn(&StreamCollection) + Send + Sync + 'static>(
&self,
f: F
) -> u64