gstreamer::structure

Struct StructureRef

pub struct StructureRef(/* private fields */);

Implementations

impl StructureRef

pub unsafe fn from_glib_borrow<'a>(ptr: *const GstStructure) -> &'a StructureRef

pub unsafe fn from_glib_borrow_mut<'a>( ptr: *mut GstStructure, ) -> &'a mut StructureRef

pub fn as_ptr(&self) -> *const GstStructure

pub fn as_mut_ptr(&self) -> *mut GstStructure

pub fn get<'a, T: FromValue<'a>>( &'a self, name: impl IntoGStr, ) -> Result<T, GetError<<<T as FromValue<'a>>::Checker as ValueTypeChecker>::Error>>

pub fn get_optional<'a, T: FromValue<'a>>( &'a self, name: impl IntoGStr, ) -> Result<Option<T>, GetError<<<T as FromValue<'a>>::Checker as ValueTypeChecker>::Error>>

pub fn value( &self, name: impl IntoGStr, ) -> Result<&SendValue, GetError<Infallible>>

pub fn get_by_quark<'a, T: FromValue<'a>>( &'a self, name: Quark, ) -> Result<T, GetError<<<T as FromValue<'a>>::Checker as ValueTypeChecker>::Error>>

pub fn get_optional_by_quark<'a, T: FromValue<'a>>( &'a self, name: Quark, ) -> Result<Option<T>, GetError<<<T as FromValue<'a>>::Checker as ValueTypeChecker>::Error>>

pub fn value_by_quark( &self, name: Quark, ) -> Result<&SendValue, GetError<Infallible>>

pub fn set(&mut self, name: impl IntoGStr, value: impl Into<Value> + Send)

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn set_if( &mut self, name: impl IntoGStr, value: impl Into<Value> + Send, predicate: bool, )

Sets field name to the given value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn set_if_some( &mut self, name: impl IntoGStr, value: Option<impl Into<Value> + Send>, )

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn set_from_iter<V: ValueType + Into<Value> + FromIterator<SendValue> + Send>( &mut self, name: impl IntoGStr, iter: impl IntoIterator<Item = impl ToSendValue>, )

Sets field name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn set_if_not_empty<V: ValueType + Into<Value> + FromIterator<SendValue> + Send>( &mut self, name: impl IntoGStr, iter: impl IntoIterator<Item = impl ToSendValue>, )

Sets field name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn set_value(&mut self, name: impl IntoGStr, value: SendValue)

Sets field name to the given value value.

Overrides any default or previously defined value for name.

pub fn set_value_if( &mut self, name: impl IntoGStr, value: SendValue, predicate: bool, )

Sets field name to the given value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn set_value_if_some( &mut self, name: impl IntoGStr, value: Option<SendValue>, )

Sets field name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn set_by_quark(&mut self, name: Quark, value: impl Into<Value> + Send)

pub fn set_by_quark_if_some( &mut self, name: Quark, value: Option<impl Into<Value> + Send>, )

pub fn set_value_by_quark(&mut self, name: Quark, value: SendValue)

pub fn set_value_by_quark_if_some( &mut self, name: Quark, value: Option<SendValue>, )

pub fn name<'a>(&self) -> &'a GStr

pub fn name_quark(&self) -> Quark

pub fn set_name(&mut self, name: impl IntoGStr)

pub fn set_name_if_some(&mut self, name: Option<impl IntoGStr>)

pub fn has_name(&self, name: &str) -> bool

pub fn has_field(&self, field: impl IntoGStr) -> bool

pub fn has_field_with_type(&self, field: impl IntoGStr, type_: Type) -> bool

pub fn has_field_by_quark(&self, field: Quark) -> bool

pub fn has_field_with_type_by_quark(&self, field: Quark, type_: Type) -> bool

pub fn remove_field(&mut self, field: impl IntoGStr)

pub fn remove_fields(&mut self, fields: impl IntoIterator<Item = impl IntoGStr>)

pub fn remove_all_fields(&mut self)

pub fn fields(&self) -> FieldIterator<'_>

pub fn iter(&self) -> Iter<'_>

pub fn nth_field_name<'a>(&self, idx: usize) -> Option<&'a GStr>

pub fn n_fields(&self) -> usize

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn can_intersect(&self, other: &StructureRef) -> bool

pub fn intersect(&self, other: &StructureRef) -> Option<Structure>

pub fn is_subset(&self, superset: &StructureRef) -> bool

pub fn fixate(&mut self)

pub fn fixate_field(&mut self, name: impl IntoGStr) -> bool

pub fn fixate_field_bool(&mut self, name: impl IntoGStr, target: bool) -> bool

pub fn fixate_field_str( &mut self, name: impl IntoGStr, target: impl IntoGStr, ) -> bool

pub fn fixate_field_nearest_double( &mut self, name: impl IntoGStr, target: f64, ) -> bool

pub fn fixate_field_nearest_fraction( &mut self, name: impl IntoGStr, target: impl Into<Fraction>, ) -> bool

pub fn fixate_field_nearest_int( &mut self, name: impl IntoGStr, target: i32, ) -> bool

pub fn foreach<F: FnMut(Quark, &Value) -> ControlFlow<()>>( &self, func: F, ) -> bool

pub fn map_in_place<F: FnMut(Quark, &mut Value) -> ControlFlow<()>>( &mut self, func: F, ) -> bool

pub fn filter_map_in_place<F: FnMut(Quark, Value) -> Option<Value>>( &mut self, func: F, )