Trait gstreamer::PluginFeatureExt

pub trait PluginFeatureExt {
    fn check_version(
        &self, 
        min_major: u32, 
        min_minor: u32, 
        min_micro: u32
    ) -> bool;

    fn get_plugin(&self) -> Option<Plugin>;

    fn get_plugin_name(&self) -> Option<String>;

    fn get_rank(&self) -> u32;

    fn load(&self) -> Option<PluginFeature>;

    fn set_rank(&self, rank: u32);
}

Trait containing all PluginFeature methods.

Implementors

DeviceProviderFactory, ElementFactory, PluginFeature, TypeFindFactory

Required Methods

fn check_version(&self, min_major: u32, min_minor: u32, min_micro: u32) -> bool

Checks whether the given plugin feature is at least the required version

min_major

minimum required major version

min_minor

minimum required minor version

min_micro

minimum required micro version

Returns

true if the plugin feature has at least the required version, otherwise false.

fn get_plugin(&self) -> Option<Plugin>

Get the plugin that provides this feature.

Returns

the plugin that provides this feature, or None. Unref with GstObjectExt::unref when no longer needed.

fn get_plugin_name(&self) -> Option<String>

Get the name of the plugin that provides this feature.

Returns

the name of the plugin that provides this feature, or None if the feature is not associated with a plugin.

fn get_rank(&self) -> u32

Gets the rank of a plugin feature.

Returns

The rank of the feature

fn load(&self) -> Option<PluginFeature>

Loads the plugin containing self if it's not already loaded. self is unaffected; use the return value instead.

Normally this function is used like this:

GstPluginFeature *loaded_feature;

loaded_feature = gst_plugin_feature_load (feature);
// presumably, we're no longer interested in the potentially-unloaded feature
gst_object_unref (feature);
feature = loaded_feature;

Returns

a reference to the loaded feature, or None on error

fn set_rank(&self, rank: u32)

Specifies a rank for a plugin feature, so that autoplugging uses the most appropriate feature.

rank

rank value - higher number means more priority rank