Struct gstreamer::value::Fraction

pub struct Fraction(pub Rational32);

Tuple Fields
0: Rational32
Implementations

impl Fraction

pub fn new(numer: i32, denom: i32) -> Self

Creates a new Ratio.

Panics

Panics if denom is zero.

pub const fn new_raw(numer: i32, denom: i32) -> Self

Creates a Fraction without checking for denom == 0 or reducing.

While this does not panic, there are several methods that will panic if used on a Fraction with denom == 0.

pub const fn from_integer(t: i32) -> Self

Creates a Fraction representing the integer t.

pub fn approximate_f32(x: f32) -> Option<Self>

pub fn approximate_f64(x: f64) -> Option<Self>

pub fn numer(&self) -> i32

pub fn denom(&self) -> i32