Crate gstreamer

Reexports

pub use miniobject::GstRc;

pub use miniobject::MiniObject;

pub use message::Message;

pub use message::MessageErrorDomain;

pub use message::MessageRef;

pub use message::MessageView;

pub use structure::Structure;

pub use structure::StructureRef;

pub use caps::Caps;

pub use caps::CapsRef;

pub use tags::Tag;

pub use tags::TagList;

pub use tags::TagListRef;

pub use buffer::Buffer;

pub use buffer::BufferMap;

pub use buffer::BufferRef;

pub use buffer::MappedBuffer;

pub use buffer::BUFFER_COPY_ALL;

pub use buffer::BUFFER_COPY_METADATA;

pub use sample::Sample;

pub use sample::SampleRef;

pub use bufferlist::BufferList;

pub use bufferlist::BufferListRef;

pub use query::Query;

pub use query::QueryRef;

pub use query::QueryView;

pub use event::Event;

pub use event::EventRef;

pub use event::EventView;

pub use event::GroupId;

pub use event::Seqnum;

pub use event::GROUP_ID_INVALID;

pub use event::SEQNUM_INVALID;

pub use context::Context;

pub use context::ContextRef;

pub use format::FormattedValue;

pub use format::GenericFormattedValue;

pub use format::SpecificFormattedValue;

pub use toc::Toc;

pub use toc::TocEntry;

pub use toc::TocEntryRef;

pub use toc::TocRef;

pub use functions::*;

Modules

buffer

bufferlist

caps

context

event

format

functions

functions

message

miniobject

prelude

query

sample

structure

tags

toc

Macros

gst_debug

gst_element_error

gst_element_info

gst_element_warning

gst_error

gst_error_msg

gst_fixme

gst_info

gst_log

gst_log_with_level

gst_memdump

gst_trace

gst_warning

Structs

Array

Bin

Bitmask

BufferCopyFlags

BufferFlags

BufferPool

BufferPoolAcquireFlags

BufferPoolAcquireParams

BufferPoolConfig

Bus

ChildProxy

Clock

ClockId

ClockTime

Continue
Continue calling the closure in the future iterations or drop it.

DateTime

DebugCategory

DebugColorFlags

DebugGraphDetails

Device

DeviceMonitor

DeviceProvider

DeviceProviderFactory

ELEMENT_METADATA_AUTHOR

ELEMENT_METADATA_DESCRIPTION

ELEMENT_METADATA_DOC_URI

ELEMENT_METADATA_ICON_NAME

ELEMENT_METADATA_KLASS

ELEMENT_METADATA_LONGNAME

Element

ElementFactory

ElementFlags

Error
A generic error capable of representing various error domains (types).

ErrorMessage

FormattedSegment

Fraction

FractionRange

GhostPad

IntRange

Iterator

List

NotifyWatchId

Object

Pad

PadLinkCheck

PadLinkSuccess

PadProbeId

PadProbeInfo

PadProbeType

PadTemplate

ParseContext

ParseFlags

Pipeline

Plugin

PluginDependencyFlags

PluginFeature

Preset

ProxyPad

Registry

SchedulingFlags

SeekFlags

SegmentFlags

SliceTypeFind

StateChangeError

StaticCaps

StaticPadTemplate

StreamFlags

SystemClock

TagSetter

TocSetter

TypeFind

TypeFindFactory

TypedValue
A statically typed Value.

URIHandler

Value
A generic value capable of carrying various types.

Enums

BufferingMode

BusSyncReply

CapsIntersectMode

ClockError

ClockReturn

ClockSuccess

ClockType

CoreError

DebugLevel

ElementMessageType

EventType

FlowError

FlowReturn

FlowSuccess

Format

IteratorError

LibraryError

PadDirection

PadLinkError

PadLinkReturn

PadMode

PadPresence

PadProbeData

PadProbeReturn

ParseError

PluginError

ProgressType

QOSType

Rank

ResourceError

SeekType

State

StateChange

StateChangeReturn

StateChangeSuccess

StreamError

StreamStatusType

StructureChangeType

TagMergeMode

TaskState

TocEntryType

TocLoopType

TocScope

Type
A GLib or GLib-based library type

TypeFindProbability

URIError

URIType

ValueOrder

Constants

BUFFER_OFFSET_NONE

CLOCK_TIME_NONE

FORMAT_PERCENT_MAX

FORMAT_PERCENT_SCALE

MSECOND

MSECOND_VAL

NSECOND

NSECOND_VAL

SECOND

SECOND_VAL

USECOND

USECOND_VAL

Traits

BinExt

BinExtManual

BufferPoolExt

BufferPoolExtManual

Cast
Upcasting and downcasting support.

ChildProxyExt

ChildProxyExtManual

ClockExt

ClockExtManual

DeviceExt

DeviceMonitorExt

DeviceProviderExt

DeviceProviderExtManual

ElementExt

ElementExtManual

GObjectExtManualGst

GhostPadExt

GstObjectExt

GstObjectExtManual

GstValueExt

IsA
Declares the "is a" relationship.

IteratorImpl

PadExt

PadExtManual

PipelineExt

PluginFeatureExt

PresetExt

ProxyPadExt

StaticType
Types that are supported by GLib dynamic typing.

SystemClockExt

TagSetterExt

TagSetterExtManual

ToValue
Converts to Value.

TocSetterExt

TypeFindImpl

URIHandlerExt

Functions

debug_bin_to_dot_data

debug_bin_to_dot_file

debug_bin_to_dot_file_with_ts

debug_get_default_threshold

debug_is_active

debug_is_colored

debug_print_stack_trace

debug_set_active

debug_set_colored

debug_set_default_threshold

debug_set_threshold_for_name

debug_set_threshold_from_string

debug_unset_threshold_for_name

init

parse_bin_from_description

parse_launch

parse_launchv

update_registry

util_get_timestamp

version

version_string

Type Definitions

ClockTimeDiff

ElementFactoryListType

Segment