Enum gstreamer::format::Signed

pub enum Signed<T> {
    Negative(T),
    Positive(T),
}

A signed wrapper.

This wrapper allows representing a signed value from a type which is originally unsigned. In C APIs, this is represented by a tuple with a signed integer positive or negative and the absolute value.

Variants

Negative(T)

Positive(T)

Implementations

impl Signed<ClockTime>

pub fn nseconds(self) -> Signed<u64>

Returns the self in nanoseconds.

pub fn from_nseconds(val: Signed<u64>) -> Self

Creates new value from nanoseconds.

pub fn useconds(self) -> Signed<u64>

Returns the self in microseconds.

pub fn from_useconds(val: Signed<u64>) -> Self

Creates new value from microseconds.

pub fn mseconds(self) -> Signed<u64>

Returns the self in milliseconds.

pub fn from_mseconds(val: Signed<u64>) -> Self

Creates new value from milliseconds.

pub fn seconds(self) -> Signed<u64>

Returns the self in seconds.

pub fn seconds_f32(self) -> f32

Returns the self in f32 seconds.

pub fn seconds_f64(self) -> f64

Returns the self in f64 seconds.

pub fn from_seconds(val: Signed<u64>) -> Self

Creates new value from seconds.

pub fn try_from_seconds_f32(seconds: f32) -> Result<Self, TryFromFloatSecsError>

Builds a new Signed<ClockTime> which value is the given number of seconds.

Returns an error if seconds is infinite or NaN, or the resulting duration in nanoseconds exceeds the u64 range.

pub fn from_seconds_f32(seconds: f32) -> Self

Builds a new Signed<ClockTime> which value is the given number of seconds.

Panics

Panics if seconds is infinite or NaN, or the resulting duration in nanoseconds exceeds the u64 range.

pub fn try_from_seconds_f64(seconds: f64) -> Result<Self, TryFromFloatSecsError>

Builds a new Signed<ClockTime> which value is the given number of seconds.

Returns an error if seconds is infinite or NaN, or the resulting duration in nanoseconds exceeds the u64 range.

pub fn from_seconds_f64(seconds: f64) -> Self

Builds a new Signed<ClockTime> which value is the given number of seconds.

Panics

Panics if seconds is infinite or NaN, or the resulting duration in nanoseconds exceeds the u64 range.

impl Signed<ClockTime>

pub fn into_inner_signed(self) -> Signed<u64>

Returns a Signed containing the inner type of self.

impl Signed<ClockTime>

pub fn signum(self) -> i32

Returns the signum for this Signed.

Returns:

0 if the number is zero.

1 if the value must be considered as positive.

-1 if the value must be considered as negative.

pub fn checked_sub(self, other: Self) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_sub_unsigned(self, other: ClockTime) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_add(self, other: Self) -> Option<Self>

Returns the checked addition self + other.

pub fn checked_add_unsigned(self, other: ClockTime) -> Option<Self>

Returns the checked addition self + other.

pub fn saturating_sub(self, other: Self) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_sub_unsigned(self, other: ClockTime) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_add(self, other: Self) -> Self

Returns the saturating addition self + other.

pub fn saturating_add_unsigned(self, other: ClockTime) -> Self

Returns the saturating addition self + other.

impl Signed<ClockTime>

fn signed_from_inner(val: u64, sign: i64) -> Option<Signed<ClockTime>>

fn positive_from_inner(val: u64) -> Option<Self>

fn negative_from_inner(val: u64) -> Option<Self>

pub fn checked_div(self, rhs: i64) -> Option<Self>

pub fn checked_div_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_rem(self, rhs: i64) -> Option<Self>

pub fn checked_rem_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_mul(self, rhs: i64) -> Option<Self>

pub fn checked_mul_unsigned(self, rhs: u64) -> Option<Self>

pub fn saturating_mul(self, rhs: i64) -> Self

pub fn saturating_mul_unsigned(self, rhs: u64) -> Self

impl Signed<Other>

pub fn into_inner_signed(self) -> Signed<u64>

Returns a Signed containing the inner type of self.

impl Signed<Other>

pub fn signum(self) -> i32

Returns the signum for this Signed.

Returns:

0 if the number is zero.

1 if the value must be considered as positive.

-1 if the value must be considered as negative.

pub fn checked_sub(self, other: Self) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_sub_unsigned(self, other: Other) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_add(self, other: Self) -> Option<Self>

Returns the checked addition self + other.

pub fn checked_add_unsigned(self, other: Other) -> Option<Self>

Returns the checked addition self + other.

pub fn saturating_sub(self, other: Self) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_sub_unsigned(self, other: Other) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_add(self, other: Self) -> Self

Returns the saturating addition self + other.

pub fn saturating_add_unsigned(self, other: Other) -> Self

Returns the saturating addition self + other.

impl Signed<Other>

fn signed_from_inner(val: u64, sign: i64) -> Option<Signed<Other>>

fn positive_from_inner(val: u64) -> Option<Self>

fn negative_from_inner(val: u64) -> Option<Self>

pub fn checked_div(self, rhs: i64) -> Option<Self>

pub fn checked_div_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_rem(self, rhs: i64) -> Option<Self>

pub fn checked_rem_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_mul(self, rhs: i64) -> Option<Self>

pub fn checked_mul_unsigned(self, rhs: u64) -> Option<Self>

pub fn saturating_mul(self, rhs: i64) -> Self

pub fn saturating_mul_unsigned(self, rhs: u64) -> Self

impl<T> Signed<T>

pub fn is_positive(self) -> bool

pub fn positive(self) -> Option<T>

Returns Some(value), where value is the inner value, if self is positive.

pub fn positive_or<E>(self, err: E) -> Result<T, E>

Transforms the Signed<T> into a Result<T, E>, mapping Positive(v) to Ok(v) and Negative(_) to Err(err).

pub fn positive_or_else<E, F: FnOnce(T) -> E>(self, err: F) -> Result<T, E>

Transforms the Signed<T> into a Result<T, E>, mapping Positive(v) to Ok(v) and Negative(v) to Err(err(v)).

pub fn is_negative(self) -> bool

pub fn negative(self) -> Option<T>

Returns Some(value), where value is the inner value, if self is negative.

pub fn negative_or<E>(self, err: E) -> Result<T, E>

Transforms the Signed<T> into a Result<T, E>, mapping Negative(v) to Ok(v) and Positive(_) to Err(err).

pub fn negative_or_else<E, F: FnOnce(T) -> E>(self, err: F) -> Result<T, E>

Transforms the Signed<T> into a Result<T, E>, mapping Negative(v) to Ok(v) and Positive(_) to Err(err(v)).

pub fn abs(self) -> T

Returns the absolute value of self.

impl<T> Signed<Option<T>>

pub fn transpose(self) -> Option<Signed<T>>

Transposes a Signed Option into an Option of a Signed.

Note that if the inner value was None, the sign is lost.

impl Signed<u64>

pub fn signum(self) -> i32

Returns the signum for this Signed.

Returns:

0 if the number is zero.

1 if the value must be considered as positive.

-1 if the value must be considered as negative.

pub fn checked_sub(self, other: Self) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_sub_unsigned(self, other: u64) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_add(self, other: Self) -> Option<Self>

Returns the checked addition self + other.

pub fn checked_add_unsigned(self, other: u64) -> Option<Self>

Returns the checked addition self + other.

pub fn saturating_sub(self, other: Self) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_sub_unsigned(self, other: u64) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_add(self, other: Self) -> Self

Returns the saturating addition self + other.

pub fn saturating_add_unsigned(self, other: u64) -> Self

Returns the saturating addition self + other.

impl Signed<u64>

fn signed_from_inner(val: u64, sign: i64) -> Option<Signed<u64>>

fn positive_from_inner(val: u64) -> Option<Self>

fn negative_from_inner(val: u64) -> Option<Self>

pub fn checked_div(self, rhs: i64) -> Option<Self>

pub fn checked_div_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_rem(self, rhs: i64) -> Option<Self>

pub fn checked_rem_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_mul(self, rhs: i64) -> Option<Self>

pub fn checked_mul_unsigned(self, rhs: u64) -> Option<Self>

pub fn saturating_mul(self, rhs: i64) -> Self

pub fn saturating_mul_unsigned(self, rhs: u64) -> Self

impl Signed<u32>

pub fn signum(self) -> i32

Returns the signum for this Signed.

Returns:

0 if the number is zero.

1 if the value must be considered as positive.

-1 if the value must be considered as negative.

pub fn checked_sub(self, other: Self) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_sub_unsigned(self, other: u32) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_add(self, other: Self) -> Option<Self>

Returns the checked addition self + other.

pub fn checked_add_unsigned(self, other: u32) -> Option<Self>

Returns the checked addition self + other.

pub fn saturating_sub(self, other: Self) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_sub_unsigned(self, other: u32) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_add(self, other: Self) -> Self

Returns the saturating addition self + other.

pub fn saturating_add_unsigned(self, other: u32) -> Self

Returns the saturating addition self + other.

impl Signed<u32>

fn signed_from_inner(val: u32, sign: i32) -> Option<Signed<u32>>

fn positive_from_inner(val: u32) -> Option<Self>

fn negative_from_inner(val: u32) -> Option<Self>

pub fn checked_div(self, rhs: i32) -> Option<Self>

pub fn checked_div_unsigned(self, rhs: u32) -> Option<Self>

pub fn checked_rem(self, rhs: i32) -> Option<Self>

pub fn checked_rem_unsigned(self, rhs: u32) -> Option<Self>

pub fn checked_mul(self, rhs: i32) -> Option<Self>

pub fn checked_mul_unsigned(self, rhs: u32) -> Option<Self>

pub fn saturating_mul(self, rhs: i32) -> Self

pub fn saturating_mul_unsigned(self, rhs: u32) -> Self

impl Signed<usize>

pub fn signum(self) -> i32

Returns the signum for this Signed.

Returns:

0 if the number is zero.

1 if the value must be considered as positive.

-1 if the value must be considered as negative.

pub fn checked_sub(self, other: Self) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_sub_unsigned(self, other: usize) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_add(self, other: Self) -> Option<Self>

Returns the checked addition self + other.

pub fn checked_add_unsigned(self, other: usize) -> Option<Self>

Returns the checked addition self + other.

pub fn saturating_sub(self, other: Self) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_sub_unsigned(self, other: usize) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_add(self, other: Self) -> Self

Returns the saturating addition self + other.

pub fn saturating_add_unsigned(self, other: usize) -> Self

Returns the saturating addition self + other.

impl Signed<usize>

fn signed_from_inner(val: usize, sign: isize) -> Option<Signed<usize>>

fn positive_from_inner(val: usize) -> Option<Self>

fn negative_from_inner(val: usize) -> Option<Self>

pub fn checked_div(self, rhs: isize) -> Option<Self>

pub fn checked_div_unsigned(self, rhs: usize) -> Option<Self>

pub fn checked_rem(self, rhs: isize) -> Option<Self>

pub fn checked_rem_unsigned(self, rhs: usize) -> Option<Self>

pub fn checked_mul(self, rhs: isize) -> Option<Self>

pub fn checked_mul_unsigned(self, rhs: usize) -> Option<Self>

pub fn saturating_mul(self, rhs: isize) -> Self

pub fn saturating_mul_unsigned(self, rhs: usize) -> Self

impl Signed<Buffers>

pub fn into_inner_signed(self) -> Signed<u64>

Returns a Signed containing the inner type of self.

impl Signed<Buffers>

pub fn signum(self) -> i32

Returns the signum for this Signed.

Returns:

0 if the number is zero.

1 if the value must be considered as positive.

-1 if the value must be considered as negative.

pub fn checked_sub(self, other: Self) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_sub_unsigned(self, other: Buffers) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_add(self, other: Self) -> Option<Self>

Returns the checked addition self + other.

pub fn checked_add_unsigned(self, other: Buffers) -> Option<Self>

Returns the checked addition self + other.

pub fn saturating_sub(self, other: Self) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_sub_unsigned(self, other: Buffers) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_add(self, other: Self) -> Self

Returns the saturating addition self + other.

pub fn saturating_add_unsigned(self, other: Buffers) -> Self

Returns the saturating addition self + other.

impl Signed<Buffers>

fn signed_from_inner(val: u64, sign: i64) -> Option<Signed<Buffers>>

fn positive_from_inner(val: u64) -> Option<Self>

fn negative_from_inner(val: u64) -> Option<Self>

pub fn checked_div(self, rhs: i64) -> Option<Self>

pub fn checked_div_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_rem(self, rhs: i64) -> Option<Self>

pub fn checked_rem_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_mul(self, rhs: i64) -> Option<Self>

pub fn checked_mul_unsigned(self, rhs: u64) -> Option<Self>

pub fn saturating_mul(self, rhs: i64) -> Self

pub fn saturating_mul_unsigned(self, rhs: u64) -> Self

impl Signed<Bytes>

pub fn into_inner_signed(self) -> Signed<u64>

Returns a Signed containing the inner type of self.

impl Signed<Bytes>

pub fn signum(self) -> i32

Returns the signum for this Signed.

Returns:

0 if the number is zero.

1 if the value must be considered as positive.

-1 if the value must be considered as negative.

pub fn checked_sub(self, other: Self) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_sub_unsigned(self, other: Bytes) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_add(self, other: Self) -> Option<Self>

Returns the checked addition self + other.

pub fn checked_add_unsigned(self, other: Bytes) -> Option<Self>

Returns the checked addition self + other.

pub fn saturating_sub(self, other: Self) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_sub_unsigned(self, other: Bytes) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_add(self, other: Self) -> Self

Returns the saturating addition self + other.

pub fn saturating_add_unsigned(self, other: Bytes) -> Self

Returns the saturating addition self + other.

impl Signed<Bytes>

fn signed_from_inner(val: u64, sign: i64) -> Option<Signed<Bytes>>

fn positive_from_inner(val: u64) -> Option<Self>

fn negative_from_inner(val: u64) -> Option<Self>

pub fn checked_div(self, rhs: i64) -> Option<Self>

pub fn checked_div_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_rem(self, rhs: i64) -> Option<Self>

pub fn checked_rem_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_mul(self, rhs: i64) -> Option<Self>

pub fn checked_mul_unsigned(self, rhs: u64) -> Option<Self>

pub fn saturating_mul(self, rhs: i64) -> Self

pub fn saturating_mul_unsigned(self, rhs: u64) -> Self

impl Signed<Default>

pub fn into_inner_signed(self) -> Signed<u64>

Returns a Signed containing the inner type of self.

impl Signed<Default>

pub fn signum(self) -> i32

Returns the signum for this Signed.

Returns:

0 if the number is zero.

1 if the value must be considered as positive.

-1 if the value must be considered as negative.

pub fn checked_sub(self, other: Self) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_sub_unsigned(self, other: Default) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_add(self, other: Self) -> Option<Self>

Returns the checked addition self + other.

pub fn checked_add_unsigned(self, other: Default) -> Option<Self>

Returns the checked addition self + other.

pub fn saturating_sub(self, other: Self) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_sub_unsigned(self, other: Default) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_add(self, other: Self) -> Self

Returns the saturating addition self + other.

pub fn saturating_add_unsigned(self, other: Default) -> Self

Returns the saturating addition self + other.

impl Signed<Default>

fn signed_from_inner(val: u64, sign: i64) -> Option<Signed<Default>>

fn positive_from_inner(val: u64) -> Option<Self>

fn negative_from_inner(val: u64) -> Option<Self>

pub fn checked_div(self, rhs: i64) -> Option<Self>

pub fn checked_div_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_rem(self, rhs: i64) -> Option<Self>

pub fn checked_rem_unsigned(self, rhs: u64) -> Option<Self>

pub fn checked_mul(self, rhs: i64) -> Option<Self>

pub fn checked_mul_unsigned(self, rhs: u64) -> Option<Self>

pub fn saturating_mul(self, rhs: i64) -> Self

pub fn saturating_mul_unsigned(self, rhs: u64) -> Self

impl Signed<Percent>

pub fn into_inner_signed(self) -> Signed<u32>

Returns a Signed containing the inner type of self.

impl Signed<Percent>

pub fn signum(self) -> i32

Returns the signum for this Signed.

Returns:

0 if the number is zero.

1 if the value must be considered as positive.

-1 if the value must be considered as negative.

pub fn checked_sub(self, other: Self) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_sub_unsigned(self, other: Percent) -> Option<Self>

Returns the checked subtraction self - other.

pub fn checked_add(self, other: Self) -> Option<Self>

Returns the checked addition self + other.

pub fn checked_add_unsigned(self, other: Percent) -> Option<Self>

Returns the checked addition self + other.

pub fn saturating_sub(self, other: Self) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_sub_unsigned(self, other: Percent) -> Self

Returns the saturating subtraction self - other.

pub fn saturating_add(self, other: Self) -> Self

Returns the saturating addition self + other.

pub fn saturating_add_unsigned(self, other: Percent) -> Self

Returns the saturating addition self + other.

impl Signed<Percent>

fn signed_from_inner(val: u32, sign: i32) -> Option<Signed<Percent>>

fn positive_from_inner(val: u32) -> Option<Self>

fn negative_from_inner(val: u32) -> Option<Self>

pub fn checked_div(self, rhs: i32) -> Option<Self>

pub fn checked_div_unsigned(self, rhs: u32) -> Option<Self>

pub fn checked_rem(self, rhs: i32) -> Option<Self>

pub fn checked_rem_unsigned(self, rhs: u32) -> Option<Self>

pub fn checked_mul(self, rhs: i32) -> Option<Self>

pub fn checked_mul_unsigned(self, rhs: u32) -> Option<Self>

pub fn saturating_mul(self, rhs: i32) -> Self

pub fn saturating_mul_unsigned(self, rhs: u32) -> Self