Trait gstreamer::format::FormattedValueNoneBuilder

pub trait FormattedValueNoneBuilder: FormattedValueFullRange {
    // Required method
    fn none() -> Self;

    // Provided method
    fn none_for_format(format: Format) -> Self { ... }
}

Required Methods

fn none() -> Self

Returns the None value for Self as a FullRange if such a value can be represented.

For SpecificFormattedValues, this results in Option::<FormattedValueIntrinsic>::None.

For GenericFormattedValue, this can only be obtained using Self::none_for_format because the None is an inner value of some of the variants.

Panics

Panics if Self is GenericFormattedValue in which case, the Format must be known.

Provided Methods

fn none_for_format(format: Format) -> Self

Returns the None value for Self if such a value can be represented.

For SpecificFormattedValues, this is the same as Self::none() if the format matches the SpecificFormattedValue’s format.

For GenericFormattedValue this is the variant for the specified format, initialized with None as a value, if the format can represent that value.

Panics

Panics if None can’t be represented by Self for format or by the requested GenericFormattedValue variant.