Module gstreamer::prelude

Re-exports

pub use meta::MetaAPI;

pub use message::MessageErrorDomain;

pub use tags::CustomTag;

pub use tags::Tag;

pub use format::FormattedValue;

pub use format::SpecificFormattedValue;

Structs

Continue
Continue calling the closure in the future iterations or drop it.

Traits

AllocatorExt

AllocatorExtManual

BufferPoolExt

BufferPoolExtManual

Cast
Upcasting and downcasting support.

ChildProxyExt

ChildProxyExtManual

ClockExt

ClockExtManual

DeviceExt

DeviceMonitorExt

DeviceMonitorExtManual

DeviceProviderExt

DeviceProviderExtManual

ElementExt

ElementExtManual

GObjectExtManualGst

GhostPadExt

GstBinExt

GstBinExtManual

GstObjectExt

GstObjectExtManual

GstParamSpecExt

GstPipelineExtManual

GstPluginExtManual

GstValueExt

IsA
Declares the "is a" relationship.

IsClassFor
Trait for mapping a class struct type to its corresponding instance type.

MulDiv
Trait for calculating val * num / denom with different rounding modes and overflow protection.

ObjectExt

ObjectType
Implemented by types representing glib::Object and subclasses of it.

PadExt

PadExtManual

PipelineExt

PluginFeatureExt

PluginFeatureExtManual

PresetExt

ProxyPadExt

StaticType
Types that are supported by GLib dynamic typing.

StaticVariantType
Returns VariantType of Self.

SystemClockExt

TagSetterExt

TagSetterExtManual

ToSendValue
Converts to SendValue.

ToValue
Converts to Value.

ToVariant
Converts to Variant.

TocSetterExt

TypeFindImpl

URIHandlerExt