Module gstreamer::prelude

Re-exports

pub use crate::memory::MemoryType;

pub use crate::message::MessageErrorDomain;

pub use crate::meta::MetaAPI;

pub use crate::meta::MetaAPIExt;

pub use crate::miniobject::IsMiniObject;

pub use crate::param_spec::GstParamSpecBuilderExt;

pub use crate::slice::Dump;

pub use crate::tags::CustomTag;

pub use crate::tags::Tag;

pub use crate::format::prelude::*;

Traits

AllocatorExt

BufferPoolExt

BufferPoolExtManual

ChildProxyExt

ChildProxyExtManual

ClockExt

ClockExtManual

ControlBindingExt

ControlBindingExtManual

ControlSourceExt

ControlSourceExtManual

DeviceExt

DeviceMonitorExt

DeviceMonitorExtManual

DeviceProviderClassExt

DeviceProviderExt

DeviceProviderExtManual

Displayable

Trait that allows accessing Display implementation on types external to this crate.

ElementClassExt

ElementExt

ElementExtManual

GObjectExtManualGst

GhostPadExt

GstBinExt

GstBinExtManual

GstObjectExt

GstObjectExtManual

GstPipelineExtManual

GstValueExt

MulDiv

Trait for calculating val * num / denom with different rounding modes and overflow protection.

OptionAdd

Trait for values and Options addition.

OptionAddAssign

Trait for values and Options addition assignment.

OptionCheckedAdd

Trait for values and Options checked addition.

OptionCheckedDiv

Trait for values and Options checked division.

OptionCheckedMul

Trait for values and Options checked multiplication.

OptionCheckedRem

Trait for values and Options checked remainder.

OptionCheckedSub

Trait for values and Options checked substraction.

OptionDiv

Trait for values and Options division.

OptionDivAssign

Trait for values and Options division assignment.

OptionEq

Trait for the equality OptionOperations.

OptionMinMax

Trait for values and Options that can be compared to get the minimum or maximum.

OptionMul

Trait for values and Options multiplication.

OptionMulAssign

Trait for values and Options multiplication assignment.

OptionOperations

Trait for inner types participating in option-operations.

OptionOrd

Trait for values and Options that can be compared for a sort-order.

OptionOverflowingAdd

Trait for values and Options overflowing addition.

OptionOverflowingDiv

Trait for values and Options overflowing division.

OptionOverflowingMul

Trait for values and Options overflowing multiplication.

OptionOverflowingRem

Trait for values and Options overflowing remainder.

OptionOverflowingSub

Trait for values and Options overflowing substraction.

OptionRem

Trait for values and Options remainder.

OptionRemAssign

Trait for values and Options remainder assignment.

OptionSaturatingAdd

Trait for values and Options saturating addition.

OptionSaturatingMul

Trait for values and Options saturating multiplication.

OptionSaturatingSub

Trait for values and Options saturating substraction.

OptionSub

Trait for values and Options substraction.

OptionSubAssign

Trait for values and Options substraction assignment.

OptionWrappingAdd

Trait for values and Options wrapping addition.

OptionWrappingDiv

Trait for values and Options wrapping division.

OptionWrappingMul

Trait for values and Options wrapping multiplication.

OptionWrappingRem

Trait for values and Options wrapping remainder.

OptionWrappingSub

Trait for values and Options wrapping substraction.

PadExt

PadExtManual

PipelineExt

PluginApiExt

PluginFeatureExt

PluginFeatureExtManual

PresetExt

ProxyPadExt

SystemClockExt

TagSetterExt

TagSetterExtManual

TaskExt

TaskHandle

A handle for a task which was pushed to a task pool.

TaskPoolExt

TaskPoolExtManual

TocSetterExt

TracerExt

TypeFindImpl

URIHandlerExt

UnixBusExtManual