Struct gstreamer::PadBuilder

pub struct PadBuilder<T> {
    pub(crate) pad: T,
    pub(crate) needs_specific_name: bool,
}

Fields
pad: Tneeds_specific_name: bool
Implementations

impl<T: IsA<GhostPad> + IsA<Pad>> PadBuilder<T>

pub fn proxy_pad_activate_function<F>(self, func: F) -> Self
where F: Fn(&ProxyPad, Option<&Object>) -> Result<(), LoggableError> + Send + Sync + 'static,

pub fn proxy_pad_activatemode_function<F>(self, func: F) -> Self
where F: Fn(&ProxyPad, Option<&Object>, PadMode, bool) -> Result<(), LoggableError> + Send + Sync + 'static,

pub fn proxy_pad_chain_function<F>(self, func: F) -> Self
where F: Fn(&ProxyPad, Option<&Object>, Buffer) -> Result<FlowSuccess, FlowError> + Send + Sync + 'static,

pub fn proxy_pad_chain_list_function<F>(self, func: F) -> Self
where F: Fn(&ProxyPad, Option<&Object>, BufferList) -> Result<FlowSuccess, FlowError> + Send + Sync + 'static,

pub fn proxy_pad_event_function<F>(self, func: F) -> Self
where F: Fn(&ProxyPad, Option<&Object>, Event) -> bool + Send + Sync + 'static,

pub fn proxy_pad_event_full_function<F>(self, func: F) -> Self
where F: Fn(&ProxyPad, Option<&Object>, Event) -> Result<FlowSuccess, FlowError> + Send + Sync + 'static,

pub fn proxy_pad_getrange_function<F>(self, func: F) -> Self
where F: Fn(&ProxyPad, Option<&Object>, u64, Option<&mut BufferRef>, u32) -> Result<PadGetRangeSuccess, FlowError> + Send + Sync + 'static,

pub fn proxy_pad_iterate_internal_links_function<F>(self, func: F) -> Self
where F: Fn(&ProxyPad, Option<&Object>) -> Iterator<Pad> + Send + Sync + 'static,

pub fn proxy_pad_link_function<F>(self, func: F) -> Self
where F: Fn(&ProxyPad, Option<&Object>, &Pad) -> Result<PadLinkSuccess, PadLinkError> + Send + Sync + 'static,

pub fn proxy_pad_query_function<F>(self, func: F) -> Self
where F: Fn(&ProxyPad, Option<&Object>, &mut QueryRef) -> bool + Send + Sync + 'static,

pub fn proxy_pad_unlink_function<F>(self, func: F) -> Self
where F: Fn(&ProxyPad, Option<&Object>) + Send + Sync + 'static,

pub fn proxy_pad_flags(self, flags: PadFlags) -> Self

pub fn with_target<P: IsA<Pad> + IsA<Object>>( self, target: &P, ) -> Result<Self, BoolError>

Specifies a target Pad for the GhostPad.

If the PadBuilder was created from a PadTemplate and the PadTemplate has a specific name_template, i.e. if it’s not a wildcard-name containing %u, %s or %d, the GhostPad will automatically be named after the name_template.

If the name_template is a wildcard-name, then the target name is used, if it is compatible. Otherwise, a specific name must be provided using PadBuilder::name or PadBuilder::maybe_name.

impl<T: IsA<Pad> + IsA<Object> + IsClass> PadBuilder<T>

pub fn new(direction: PadDirection) -> Self

Creates a PadBuilder with the specified PadDirection.

An automatically generated name will be assigned. Use PadBuilder::name or PadBuilder::maybe_name to define a specific name.

pub fn from_static_template(templ: &StaticPadTemplate) -> Self

Creates a PadBuilder from the specified StaticPadTemplate.

If the StaticPadTemplate has a specific name_template, i.e. if it’s not a wildcard-name containing %u, %s or %d, the Pad will automatically be named after the name_template.

Use PadBuilder::name or PadBuilder::maybe_name to specify a different name.

pub fn from_template(templ: &PadTemplate) -> Self

Creates a PadBuilder from the specified PadTemplate.

If the PadTemplate has a specific name_template, i.e. if it’s not a wildcard-name containing %u, %s or %d, the Pad will automatically be named after the name_template.

Use PadBuilder::name or PadBuilder::maybe_name to specify a different name.

pub fn name(self, name: impl IntoGStr) -> Self

Sets the name of the Pad.

pub fn maybe_name<N: IntoGStr>(self, name: Option<N>) -> Self

Optionally sets the name of the Pad.

This method is convenient when the name is provided as an Option. If the name is None, this has no effect.

pub fn activate_function<F>(self, func: F) -> Self
where F: Fn(&T, Option<&Object>) -> Result<(), LoggableError> + Send + Sync + 'static,

pub fn activatemode_function<F>(self, func: F) -> Self
where F: Fn(&T, Option<&Object>, PadMode, bool) -> Result<(), LoggableError> + Send + Sync + 'static,

pub fn chain_function<F>(self, func: F) -> Self
where F: Fn(&T, Option<&Object>, Buffer) -> Result<FlowSuccess, FlowError> + Send + Sync + 'static,

pub fn chain_list_function<F>(self, func: F) -> Self
where F: Fn(&T, Option<&Object>, BufferList) -> Result<FlowSuccess, FlowError> + Send + Sync + 'static,

pub fn event_function<F>(self, func: F) -> Self
where F: Fn(&T, Option<&Object>, Event) -> bool + Send + Sync + 'static,

pub fn event_full_function<F>(self, func: F) -> Self
where F: Fn(&T, Option<&Object>, Event) -> Result<FlowSuccess, FlowError> + Send + Sync + 'static,

pub fn getrange_function<F>(self, func: F) -> Self
where F: Fn(&T, Option<&Object>, u64, Option<&mut BufferRef>, u32) -> Result<PadGetRangeSuccess, FlowError> + Send + Sync + 'static,

pub fn iterate_internal_links_function<F>(self, func: F) -> Self
where F: Fn(&T, Option<&Object>) -> Iterator<Pad> + Send + Sync + 'static,

pub fn link_function<F>(self, func: F) -> Self
where F: Fn(&T, Option<&Object>, &Pad) -> Result<PadLinkSuccess, PadLinkError> + Send + Sync + 'static,

pub fn query_function<F>(self, func: F) -> Self
where F: Fn(&T, Option<&Object>, &mut QueryRef) -> bool + Send + Sync + 'static,

pub fn unlink_function<F>(self, func: F) -> Self
where F: Fn(&T, Option<&Object>) + Send + Sync + 'static,

pub fn flags(self, flags: PadFlags) -> Self

pub fn build(self) -> T

Builds the Pad.

Panics

Panics if the Pad was built from a PadTemplate with a wildcard-name name_template (i.e. containing %u, %s or %d) and no specific name was provided using PadBuilder::name or PadBuilder::maybe_name, or for GhostPads, by defining a target.